SDiagonalizability

Documentation for SDiagonalizability.

SDiagonalizability._assert_matrix_is_undirected_laplacian

SDiagonalizability._extract_independent_cols

SDiagonalizability._find_indices_1neg

SDiagonalizability._laplacian_spectra_01neg

SDiagonalizability._pot_kernel_eigvecs_01neg

SDiagonalizability._pot_nonkernel_eigvecs_01neg

SDiagonalizability._rank_rtol

SDiagonalizability._rank_rtol

SDiagonalizability._typed_laplacian_spectra_01neg

SDiagonalizability._typed_laplacian_spectra_01neg

SDiagonalizability._typed_laplacian_spectra_01neg

SDiagonalizability._typed_laplacian_spectra_01neg

SDiagonalizability._typed_laplacian_spectra_01neg

SDiagonalizability.check_spectrum_integrality

SDiagonalizability._assert_matrix_is_undirected_laplacian — Method

_assert_matrix_is_undirected_laplacian(L)

Validate that L is the Laplacian matrix of an undirected, possibly weighted graph.

Arguments

L::AbstractMatrix{<:Integer}: the purported Laplacian matrix.

Returns

nothing: if the check is passed, no output is produced.

Throws

DomainError: if L is not symmetric or has nonzero row sums.

Examples

The Laplacian matrix of the (undirected) star graph on 5 vertices passes the check:

julia> L = [ 4  -1  -1  -1  -1;
            -1   1   0   0   0;
            -1   0   1   0   0;
            -1   0   0   1   0;
            -1   0   0   0   1]
5×5 Matrix{Int64}:
  4  -1  -1  -1  -1
 -1   1   0   0   0
 -1   0   1   0   0
 -1   0   0   1   0
 -1   0   0   0   1

julia> isnothing(SDiagonalizability._assert_matrix_is_undirected_laplacian(L))
true

The adjacency matrix of the (undirected) cycle graph on 4 vertices is symmetric but has nonzero row sums, so it fails the check:

julia> A = [0  1  0  1;
            1  0  1  0;
            0  1  0  1;
            1  0  1  0]
4×4 Matrix{Int64}:
 0  1  0  1
 1  0  1  0
 0  1  0  1
 1  0  1  0

julia> SDiagonalizability._assert_matrix_is_undirected_laplacian(A)
ERROR: DomainError with [0 1 0 1; 1 0 1 0; 0 1 0 1; 1 0 1 0]:
Matrix has nonzero row sums; cannot be an (undirected) Laplacian
[...]

Both the in-degree and out-degree Laplacian matrices of this random tournament digraph have zero row sums but are not symmetric, so they fail the check. (These are the two standard ways of extending the concept of the Laplacian to directed graphs [1, p. 196].)

julia> G = random_tournament_digraph(3; seed=87)
{3, 3} directed simple Int64 graph

julia> L_in = laplacian_matrix(G; dir=:in)
3×3 SparseArrays.SparseMatrixCSC{Int64, Int64} with 5 stored entries:
  ⋅  ⋅   ⋅
 -1  2  -1
 -1  ⋅   1

julia> SDiagonalizability._assert_matrix_is_undirected_laplacian(L_in)
ERROR: DomainError with sparse([2, 3, 2, 2, 3], [1, 1, 2, 3, 3], [-1, -1, 2, -1, 1], 3, 3):
Matrix is not symmetric; cannot be an (undirected) Laplacian
[...]

julia> L_out = laplacian_matrix(G; dir=:out)
3×3 SparseArrays.SparseMatrixCSC{Int64, Int64} with 5 stored entries:
 2  -1  -1
 ⋅   ⋅   ⋅
 ⋅  -1   1

julia> SDiagonalizability._assert_matrix_is_undirected_laplacian(L_out)
ERROR: DomainError with sparse([1, 1, 3, 1, 3], [1, 2, 2, 3, 3], [2, -1, -1, -1, 1], 3, 3):
Matrix is not symmetric; cannot be an (undirected) Laplacian
[...]

Notes

If edges are to be bidirectional, then L must be symmetric. L must also have zero row sums, since the (i, i)-th entry is the weighted degree of node i (the sum of all incident edges' weights) and the (i, j)-th entry for i ≠ j is the negation of the weight of edge (i, j) (or simply 0, if no such edge exists).

Given the highly optimized, lazy, zero-allocation implementation of LinearAlgebra.issymmetric, the symmetry check is performed first. (Both steps are O(n²) in the worst case, but testing for symmetry is far more performant in practice.) This also allows us to (also lazily) check for nonzero column sums rather than nonzero row sums (since these are equivalent for symmetric matrices) in the second step, taking advantage of Julia's column-major storage model.

At first blush, it may seem as though the choice of DomainError over something like ArgumentError (or even simply the return of a boolean) constitutes poor design. However, this is informed by the ad hoc use of assert_matrix_is_undirected_laplacian to validate inputs for other functions requiring Laplacian matrices. This function is never meant to be publicly exposed on its own.

SDiagonalizability._extract_independent_cols — Method

_extract_independent_cols(A)

Return a (not necessarily unique) independent spanning subset of the columns of A.

Computing a QR decomposition of A, the scaling coefficients from the orthogonalization process are used to determine the rank (rather than recompute it with an SVD), while the pivots are used to extract a spanning set of independent columns.

Arguments

A::AbstractMatrix{T<:Integer}: the matrix whose independent columns to extract.

Returns

::AbstractMatrix{T}: a spanning set of independent columns of A.

Examples

Observe how columns with greater Euclidean norms are given priority in the pivot ordering:

julia> A = [3  0  0  0  2  1   5   0
            0  3  0  0  2  1  -5   0
            0  0  3  0  2  1   5   4
            0  0  0  3  2  1   0  -4
            0  0  0  0  0  0   0   0]
5×8 Matrix{Int64}:
 3  0  0  0  2  1   5   0
 0  3  0  0  2  1  -5   0
 0  0  3  0  2  1   5   4
 0  0  0  3  2  1   0  -4
 0  0  0  0  0  0   0   0

julia> SDiagonalizability._extract_independent_cols(A)
5×4 Matrix{Int64}:
  5   0  2  3
 -5   0  2  0
  5   4  2  0
  0  -4  2  0
  0   0  0  0

Notes

TODO: Discuss unusual use of _rank_rtol

SDiagonalizability._find_indices_1neg — Method

_find_indices_1neg(eigvecs_01neg)

Find the indices of {-1,1}-eigenvectors given a map of eigenspaces.

More precisely, given a map from each eigenvalue to all {-1,0,1}-vectors in the associated eigenspace, this function returns a map from each eigenvalue to the indices of the {-1,1}-vectors in each value from the input map.

Arguments

eigvecs_01neg::AbstractDict{Int,<:AbstractMatrix{Int}}: a map from eigenvalues to eigenspaces, TODO: Write here

Returns

::AbstractDict{Int,Vector{Int}}: TODO: Write here

Examples

TODO: Write here

Notes

TODO: Justify decision to use Vector{Int} rather than BitVector for the indices

SDiagonalizability._laplacian_spectra_01neg — Method

_laplacian_spectra_01neg(L::AbstractMatrix{<:Integer})

TODO: Write here

Arguments

L::AbstractMatrix{<:Integer}: the Laplacian matrix on whose {-1,0,1}-spectrum we are to compute data.

Returns

::_LaplacianSpectrum01Neg: a struct containing the following fields:

laplacian_matrix::Matrix{Int}: a (casted) copy of L, avoiding shared mutability.

diagonalizable_01neg::Bool: whether the Laplacian is {-1,0,1}-diagonalizable.

diagonalizable_1neg::Bool: whether the Laplacian is {-1,1}-diagonalizable.

eigspaces_01neg::Union{Nothing,OrderedDict{Int,_Eigenspace01Neg}}: a map from eigenvalues to eigenspaces, sorted first by ascending multiplicity then by ascending eigenvalue. (This field is nothing if the Laplacian is not {-1,0,1}-diagonalizable.)

Examples

TODO: Write here

Notes

TODO: Write here

SDiagonalizability._pot_kernel_eigvecs_01neg — Method

_pot_kernel_eigvecs_01neg(n)

Lazily generate all potential kernel {-1,0,1}-eigenvectors of an order n Laplacian.

Each vector is normalized so that its first nonzero entry is 1, enforcing pairwise linear independence between all generated vectors.

Arguments

n::Integer: the order of the Laplacian matrix of some undirected graph for which to find potential kernel eigenvectors.

Returns

eigvec_generator::Iterators.Flatten{<:Base.Generator}: a lazily evaluated iterator over all {-1,0,1}-vectors in ℝⁿ, unique up to span.

Throws

DomainError: if n is negative.

Examples

Generate all potential kernel eigenvectors for an order 2 Laplacian matrix:

julia> hcat(SDiagonalizability._pot_kernel_eigvecs_01neg(3)...)
3×13 Matrix{Int64}:
  1   1   1   1  1  1   1  1  1   0  0  0  0
 -1  -1  -1   0  0  0   1  1  1   1  1  1  0
 -1   0   1  -1  0  1  -1  0  1  -1  0  1  1

Notes

The number of potential kernel eigenvectors (unique up to span) for an order n Laplacian matrix is given by (3ⁿ - 1) / 2. See also the relevant OEIS sequence [2].

Regrettably, the implementation here is rather clunky and unidiomatic, but it is worth noting that eigenvector generation (and subsequent processing by _laplacian_spectra_01neg) is one of two major bottlenecks in the overall S-bandwidth minimization algorithm. Given how much potential there is for optimization in this piece of code, we thus prioritize performance over readability in this particular case, making every effort to include inline comments wherever clarification may be needed.

SDiagonalizability._pot_nonkernel_eigvecs_01neg — Method

_pot_nonkernel_eigvecs_01neg(n)

Lazily generate all potential non-kernel {-1,0,1}-eigenvectors of an order n Laplacian.

Each vector is normalized so that its first nonzero entry is 1, enforcing pairwise linear independence between all generated vectors. Since all Laplacian matrices have pairwise orthogonal eigenspaces and the all-ones vector is always in the kernel, every non-kernel {-1,0,1}-eigenvector must also have an equal number of -1's and 1's.

Arguments

n::Integer: the order of the Laplacian matrix of some undirected graph for which to find potential non-kernel eigenvectors.

Returns

eigvec_generator::Iterators.Flatten{<:Base.Generator}: a lazily evaluated iterator over all {-1,0,1}-vectors in ℝⁿ orthogonal to the all-ones kernel vector, unique up to span.

Throws

DomainError: if n is negative.

Examples

Generate all potential non-kernel eigenvectors of an order 4 Laplacian matrix:

julia> hcat(SDiagonalizability._pot_nonkernel_eigvecs_01neg(4)...)
4×9 Matrix{Int64}:
  1   1   1   1   1   1   0   0   0
 -1   0   0  -1  -1   1   1   1   0
  0  -1   0  -1   1  -1  -1   0   1
  0   0  -1   1  -1  -1   0  -1  -1

Notes

The number of potential non-kernel eigenvectors (unique up to span) for an order n Laplacian matrix is, by non-trivial combinatorial arguments, equal to the number of humps in all Motzkin paths of length n. See also the relevant OEIS sequence [3].

Regrettably, the implementation here is rather clunky and unidiomatic, but it is worth noting that eigenvector generation (and subsequent processing by _laplacian_spectra_01neg) is one of two major bottlenecks in the overall S-bandwidth minimization algorithm. Given how much potential there is for optimization in this piece of code, we thus prioritize performance over readability in this particular case, making every effort to include inline comments wherever clarification may be needed.

SDiagonalizability._rank_rtol — Method

function _rank_rtol(A::AbstractMatrix{<:Real})

Return a reasonable relative tolerance for computing matrix rank via SVD or QRD.

If using SVD to compute rank with singular values Ω, then the intended use case is to numerically estimate the rank r of A by r = count(Ω .> maximum(Ω) * _rank_rtol(A)). Similarly, if using QRD to compute rank with scaling coefficients Δ (the diagonal entries of the R matrix in A = QR), then the user should numerically estimate the rank by r = count(Δ .> maximum(Δ) * _rank_rtol(A)).

Since the machine epsilon is only defined for floating-point types, we call eps() (defaulting to Float64) instead of eps(eltype(A)) when computing the tolerance. If the elements of A are floats, on the other hand, then _rank_rtol dispatches to a separate method that uses the element type's machine epsilon.

Arguments

A::AbstractMatrix{<:Real}: the matrix for which to compute a tolerance.

Returns

tol::Float64: a reasonable relative tolerance for computing matrix rank via SVD or QRD. This scales proportionally to the maximum dimension of A.

Notes

This is very much an ad hoc function applied specifically for numerical stability within this package and is never meant to be used as part of the public API. The sqrt(maximum(size(A)) * eps()) formula is inspired by NumPy's default relative tolerance of maximum(size(A)) * eps() in the numpy.linalg.matrix_rank function [4], with a square root taken for additional robustness. (In particular, the short, fat matrices holding all {-1,0,1}-eigenvectors of a Laplacian matrix are often ill-conditioned, hence the need for this higher tolerance.)

In the case of using SVD to compute matrix rank, the decision to default to the Float64 machine epsilon is motivated by how LinearAlgebra.rank handles non-floating-point matrices, as LAPACK automatically converts to Float64 under the hood. In the case of QRD, our ad hoc use of this function in SDiagonalizability._extract_independent_cols takes in floating-point matrices anyway, so this particular method is no longer relevant.

SDiagonalizability._rank_rtol — Method

function _rank_rtol(A::AbstractMatrix{<:AbstractFloat})

Return a reasonable relative tolerance for computing matrix rank via SVD or QRD.

If using SVD to compute rank with singular values Ω, then the intended use case is to numerically estimate the rank r of A by r = count(Ω .> maximum(Ω) * _rank_rtol(A)). Similarly, if using QRD to compute rank with scaling coefficients Δ (the diagonal entries of the R matrix in A = QR), then the user should numerically estimate the rank by r = count(Δ .> maximum(Δ) * _rank_rtol(A)).

Arguments

A::AbstractMatrix{T<:AbstractFloat}: the matrix for which to compute a tolerance.

Returns

tol::Float64: a reasonable relative tolerance for computing matrix rank via SVD or QRD. This scales proportionally to the maximum dimension of A and the machine epsilon of the element type T.

Notes

This is very much an ad hoc function applied specifically for numerical stability within this package and is never meant to be used as part of the public API. The sqrt(maximum(size(A)) * eps()) formula is inspired by NumPy's default relative tolerance of maximum(size(A)) * eps() in the numpy.linalg.matrix_rank function [4], with a square root taken for additional robustness. (In particular, the short, fat matrices holding all {-1,0,1}-eigenvectors of a Laplacian matrix are often ill-conditioned, hence the need for this higher tolerance.)

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_ArbitraryGraphLaplacian)

TODO: Write here

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_CompleteGraphLaplacian)

TODO: Write here

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_EmptyGraphLaplacian)

TODO: Write here

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_NullGraphLaplacian)

TODO: Write here

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_TypedLaplacian)

TODO: Write here

SDiagonalizability.check_spectrum_integrality — Method

check_spectrum_integrality(A)

Check whether the eigenvalues of A are integers (up to floating-point error).

If the eigenvalues are indeed all integers, then an eigenvalue-multiplicity map is constructed as well.

Arguments

A::AbstractMatrix{<:Integer}: the matrix whose eigenvalues to check for integrality.

Returns

::SpectrumIntegralResult: a struct containing the following fields:

matrix::Matrix{Int}: a (casted) copy of A, avoiding shared mutability.

spectrum_integral::Bool: whether the eigenvalues of A are integers.

multiplicities::Union{Nothing,OrderedDict{Int,Int}}: a map from eigenvalues to multiplicities, sorted first by ascending multiplicity then by ascending eigenvalue. (This field is nothing if the eigenvalues are not all integers.)

Examples

Confirm that the rotation matrix by π/2 radians counterclockwise is not spectrum integral (rather, it has eigenvalues ±i [5, p. 1]):

julia> R = Int8.([0 -1; 1 0])
2×2 Matrix{Int8}:
 0  -1
 1   0

julia> res = check_spectrum_integrality(R);

julia> res.matrix
2×2 Matrix{Int64}:
 0  -1
 1   0

julia> res.spectrum_integral
false

julia> isnothing(res.multiplicities)
true

Confirm that the adjacency matrix of the Petersen graph is spectrum integral, with correct eigenvalues and multiplicities of {3: 1, -2: 4, 1: 5} [6, p. 2]:

julia> G = smallgraph(:petersen)
{10, 15} undirected simple Int64 graph

julia> A = adjacency_matrix(G)
10×10 SparseArrays.SparseMatrixCSC{Int64, Int64} with 30 stored entries:
 ⋅  1  ⋅  ⋅  1  1  ⋅  ⋅  ⋅  ⋅
 1  ⋅  1  ⋅  ⋅  ⋅  1  ⋅  ⋅  ⋅
 ⋅  1  ⋅  1  ⋅  ⋅  ⋅  1  ⋅  ⋅
 ⋅  ⋅  1  ⋅  1  ⋅  ⋅  ⋅  1  ⋅
 1  ⋅  ⋅  1  ⋅  ⋅  ⋅  ⋅  ⋅  1
 1  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  1  1  ⋅
 ⋅  1  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  1  1
 ⋅  ⋅  1  ⋅  ⋅  1  ⋅  ⋅  ⋅  1
 ⋅  ⋅  ⋅  1  ⋅  1  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  ⋅  1  ⋅  1  1  ⋅  ⋅

julia> res = check_spectrum_integrality(A);

julia> res.matrix
10×10 Matrix{Int64}:
 0  1  0  0  1  1  0  0  0  0
 1  0  1  0  0  0  1  0  0  0
 0  1  0  1  0  0  0  1  0  0
 0  0  1  0  1  0  0  0  1  0
 1  0  0  1  0  0  0  0  0  1
 1  0  0  0  0  0  0  1  1  0
 0  1  0  0  0  0  0  0  1  1
 0  0  1  0  0  1  0  0  0  1
 0  0  0  1  0  1  1  0  0  0
 0  0  0  0  1  0  1  1  0  0

julia> res.spectrum_integral
true

julia> res.multiplicities
OrderedCollections.OrderedDict{Int64, Int64} with 3 entries:
  3  => 1
  -2 => 4
  1  => 5

Notes

If an undirected graph with integer edge weights is {-1,0,1}-diagonalizable (or, more restrictively, {-1,1}-diagonalizable), then its Laplacian matrix has integer eigenvalues [7, p. 300]. Hence, validating Laplacian integrality serves as a useful screening step in SDiagonalizability.jl's principal S-bandwidth minimzation algorithm.

It is, perhaps, an odd choice to sort the eigenvalue/multiplicity pairs in this file rather than in src/s_bandwidth.jl—after all, in the context of the overarching S-bandwidth algorithm, this ordering is only ever used to determine which eigenspaces are searched for S-bases first. However, the inclusion of check_spectrum_integrality in the public API motivates the enforcement of a consistent, natural ordering of the multiplicites map in each (potentially user-facing) SpectrumIntegralResult instance.

Of course, we make it a point to still sort multiplicities as desired (first by ascending multiplicity then by ascending eigenvalue) in src/s_bandwidth.jl itself whenever needed by another function; this seems more robust than relying on check_spectrum_integrality to do so or even folding the logic into the SpectrumIntegralResult inner constructor.

TODO: Should some of this explanation go in the SpectrumIntegralResult docstring instead?

[1]

J. J. P. Veerman and R. Lyons. A Primer on Laplacian Dynamics in Directed Graphs. Nonlinear Phenomena in Complex Systems 23, 196–206 (2020).

[2]

N. J. A. Sloane, a(n) = (3^n - 1)/2. Entry A003462 in The On-Line Encyclopedia of Integer Sequences (2025). Accessed: 2025-05-22.

[3]

E. Deutsch. Number of humps in all Motzkin paths of length n. Entry A097861 in The On-Line Encyclopedia of Integer Sequences (2025). Accessed: 2025-05-22.

[4]

NumPy Developers, numpy.linalg.matrix_rank. NumPy reference documentation (stable) (2025). Accessed: 2025-05-22.

[5]

D. Joyce. Rotations and complex eigenvalues. Lecture notes, Math 130: Linear Algebra, Clark University (2015).

[6]

J. Fox. Lecture 19: The Petersen graph and Moore graphs. Lecture notes, MAT 307: Combinatorics, Massachusetts Institute of Technology (2009).

[7]

N. Johnston and S. Plosker. Laplacian {−1,0,1}- and {−1,1}-diagonalizable graphs. Linear Algebra and its Applications 704, 309–39 (2025).