SDiagonalizability.jl – Private API

Documentation for SDiagonalizability's private API.

Note

The following documentation covers only the private API of the package. For public details, see the public API documentation.

Warning

The following bindings are internal; they may change or be removed in future versions.

SDiagonalizability._ORTHOGONALITY — Constant

_ORTHOGONALITY::_Orthogonality

A singleton instance of the semantically unique _Orthogonality type.

This constant is used to represent the property of pairwise orthogonality for a collection of vectors and should always be used in favor of instantiating a new object with _Orthogonality() every time.

Notes

Simply by definition, we are certain that the k property of any _Orthogonality instance will always be 1, so given that the struct is immutable and we do not intend to add any other distinguishing fields/properties, it is idiomatic to define this sort of singleton instance for universal use. This design choice also avoids unnecessary allocations (although any overhead will be negligible in practice).

SDiagonalizability._QUASI_ORTHOGONALITY — Constant

_QUASI_ORTHOGONALITY::_QuasiOrthogonality

A singleton instance of the semantically unique _QuasiOrthogonality type.

This constant is used to represent the property of quasi-orthogonality for a collection of vectors and should always be used in favor of instantiating a new object with _QuasiOrthogonality() every time.

Notes

Simply by definition, we are certain that the k property of any _QuasiOrthogonality instance will always be 2, so given that the struct is immutable and we do not intend to add any other distinguishing fields/properties, it is idiomatic to define this sort of singleton instance for universal use. This design choice also avoids unnecessary allocations (although any overhead will be negligible in practice).

SDiagonalizability._ArbitraryGraphLaplacian — Type

struct _ArbitraryGraphLaplacian

A wrapper for the Laplacian of any graph on n ≥ 3 nodes with no particular structure of interest.

Fields

matrix::Matrix{Int}: the n-by-n Laplacian in question. (Necessarily symmetric with zero row sums for some n ≥ 3.)

Supertype Hierarchy

_ArbitraryGraphLaplacian <: _TypedLaplacian <: Any

Notes

The only graph on zero nodes is the null graph (handled by the _NullGraphLaplacian type), the only graph on one node is an empty graph (handled by the _EmptyGraphLaplacian type), and the only graphs on two nodes are an empty graph and a complete graph (handled by the _CompleteGraphLaplacian type), so we are confident in the assumption that n ≥ 3 for any _ArbitraryGraphLaplacian instance.

See the _typed_laplacian_spectra_01neg documentation for more details on the {-1,0,1}-spectral properties of this Laplacian type.

See also the documentation for parent type _TypedLaplacian if seeking justification for typing the matrix field as Matrix{Int} rather than AbstractMatrix{<:Integer}.

SDiagonalizability._CompleteGraphLaplacian — Type

struct _CompleteGraphLaplacian

A wrapper for the Laplacian of any uniformly weighted complete graph on n ≥ 2 nodes.

Fields

matrix::Matrix{Int}: the n-by-n Laplacian in question. (Necessarily symmetric with zero row sums and all off-diagonal entries equal to -weight for some n ≥ 2.)

weight::Int: the weight of each edge in the graph represented by matrix. (Necessarily a nonzero integer.)

Supertype Hierarchy

_CompleteGraphLaplacian <: _TypedLaplacian <: Any

Notes

The only graph on zero nodes is the null graph (handled by the _NullGraphLaplacian type) and the only graph on one node is an empty graph (handled by the _EmptyGraphLaplacian type), so we are confident in the assumption that n ≥ 2 for any _CompleteGraphLaplacian instance.

See the _typed_laplacian_spectra_01neg documentation for more details on the {-1,0,1}-spectral properties of this Laplacian type.

See also the documentation for parent type _TypedLaplacian if seeking justification for typing the matrix field as Matrix{Int} rather than AbstractMatrix{<:Integer}.

SDiagonalizability._Eigenspace01Neg — Type

struct _Eigenspace01Neg

Data on an eigenspace of some Laplacian matrix with respect to the {-1,0,1}-spectrum.

Fields

multiplicity::Int: the multiplicity of the eigenspace.

eigvecs_01neg::AbstractMatrix{Int}: the {-1,0,1}-eigenvectors in this eigenspace, stored as columns of a matrix.

indices_1neg::Vector{Int}: the indices of all {-1,1}-columns in eigvecs_01neg.

basis_01neg::Union{Nothing,Matrix{Int}}: a {-1,0,1}-basis for the eigenspace, if one exists. (This field is nothing if no such basis exists.)

basis_1neg::Union{Nothing,Matrix{Int}}: a {-1,1}-basis for the eigenspace, if one exists. (This field is nothing if no such basis exists.)

Notes

This helper struct is, above all else, used for the sake of efficient data storage in user-facing structs, with the LaplacianSpectrum01Neg constructor flattening data stored across several _Eigenspace01Neg instances into top-level fields. This justifies the exclusion of the eigval field, which is redundant in the context of LaplacianSpectrum01Neg (the keys of the eigspaces_01neg::{Int,_Eigenspace01Neg} map already specify the eigenvalues).

See the _find_indices_1neg documentation if seeking justification for why the indices_1neg field is a Vector{Int} rather than a BitVector. (In short, very few columns have entries exclusively from {-1,1}, so Vector{Int}'s consume less memory.)

SDiagonalizability._EmptyGraphLaplacian — Type

struct _EmptyGraphLaplacian

A wrapper for the Laplacian matrix of any graph with no edges on n ≥ 1 nodes.

Fields

matrix::Matrix{Int}: the n-by-n Laplacian in question. (Necessarily equal to zeros(Int, n, n) for some n ≥ 1.)

Supertype Hierarchy

_EmptyGraphLaplacian <: _TypedLaplacian <: Any

Notes

We are confident in the assumption that n is at least 1 because the only graph on zero nodes is the null graph, which is handled by the _NullGraphLaplacian type.

See the _typed_laplacian_spectra_01neg documentation for more details on the {-1,0,1}-spectral properties of this Laplacian type.

See also the documentation for parent type _TypedLaplacian if seeking justification for typing the matrix field as Matrix{Int} rather than AbstractMatrix{<:Integer}.

SDiagonalizability._KOrthogonality — Type

abstract type _KOrthogonality

An abstract type representing the property of k-orthogonality of a collection of vectors.

Recall that an (ordered) collection of vectors v₁, v₂, ..., vₙ is said to be k-orthogonal if we have the inner product ⟨vᵢ, vⱼ⟩ = 0 whenever |i - j| ≥ k (i.e., if every pair of vectors at least k indices apart is orthogonal). This is equivalent to the vectors' Gram matrix having bandwidth at most k, where we define the bandwidth of a matrix A to be the minimum integer k ≥ 1 such that aᵢⱼ = 0 whenever |i - j| ≥ k [JP25, p. 313]. (Note that many texts instead define matrix bandwidth using zero-based indexing—that is, with the condition |i - j| > k [Maf14, p. 186].)

This type is used as a template for concretely defined properties corresponding to specific values of k. For each family of values, we apply a separate algorithm to determine whether there exists a permutation of a given collection of vectors that induces k-orthogonality.

Interface

Concrete subtypes of _KOrthogonality must implement the following properties:

k::Int: the k-orthogonality parameter.

k need not necessarily be a field of the subtype (especially in the case of single-value types like _Orthogonality and _QuasiOrthogonality, in which case making k a field may erroneously imply that it is variable), but it must be accessible via Base.getproperty.

Subtypes

_Orthogonality: k-orthogonality for k = 1.

_QuasiOrthogonality: k-orthogonality for k = 2.

_WeakOrthogonality: k-orthogonality for k > 2.

Notes

Given the domain restriction of k to the positive integers, combined with the relatively small nature (≤ 25) of bandwidths examined by this package's principal S-bandwidth minimization algorithm, it may seem prudent to ask why we do not type the k field as UInt8 instead of Int. An equally compelling argument can be made to type it as Integer for genericity as well. However, many methods (both within this module and externally) often expect Int-typed parameters, so we settle on Int as our field type. However, in any constructors/factories of _KOrthogonality subtypes, we allow the k-orthogonality input parameter to be any generic Integer, simply casting it to an Int later on.

On that note, the stipulation in the interface contract that k be an accessible property even for semantically unique types like _Orthogonality and _QuasiOrthogonality may seem somewhat odd, as it will never accessed in practice by _find_k_orthogonal_basis. (Rather, the type itself—which already implicitly encodes the value of k—directly informs the choice of algorithm in said function.) However, we require it nonetheless for the sake of consistency future extensibility, just in case the API is later modified to require more generic handling of _KOrthogonality instances whose specific subtypes are not known based on their k-orthogonality parameter.

Another design choice worth scrutinizing is the "_WeakOrthogonality" name. "Orthogonality" is a term as old as time, and "quasi-orthogonality" too has risen to prominence in recent years (e.g., [JP25, p. 313]), but "weak orthogonality," strictly speaking, is not standard terminology. We have taken the liberty of coining it here for k > 2 to emphasize its weaker nature compared to orthogonality and quasi-orthogonality. That said, we make no attempts to formally introduce as a new definition in the broader literature beyond its ad hoc use (and but internally, at that) in this module.

Finally, we note that it may be slightly misleading to state that this type is used to inform our choice of algorithm in determining whether a collection of vectors is k-orthogonalizable (i.e., whether there exists a k-orthogonal permutation of said collection). Rather, we use it when determining whether a large, linearly dependent set of vectors contains an k-orthogonal independent spanning subset. This is still related to the long-standing matrix bandwidth minimization problem, but it differs in that we can take advantage of dynamic programming techniques when recursively searching for such a subset. (See the _find_k_orthogonal_basis documentation for more details.)

SDiagonalizability._NullGraphLaplacian — Type

struct _NullGraphLaplacian

A wrapper for the Laplacian matrix of the (unique) graph with no nodes.

Fields

matrix::Matrix{Int}: the Laplacian in question. (Necessarily a 0×0 Matrix{Int}.)

Supertype Hierarchy

_NullGraphLaplacian <: _TypedLaplacian <: Any

Notes

See the _typed_laplacian_spectra_01neg documentation for more details on the {-1,0,1}-spectral properties of this Laplacian type.

See also the documentation for parent type _TypedLaplacian if seeking justification for typing the matrix field as Matrix{Int} rather than AbstractMatrix{<:Integer}.

SDiagonalizability._Orthogonality — Type

struct _Orthogonality

The property of pairwise orthogonality for a collection of vectors.

Recall that a collection of vectors v₁, v₂, ..., vₙ is said to be pairwise orthogonal if we have the inner product ⟨vᵢ, vⱼ⟩ = 0 whenever |i - j| ≥ 1 (i.e., if every pair of vectors is orthogonal). This is equivalent to the vectors' Gram matrix being diagonal.

Properties

k::Int: the k-orthogonality parameter. (Necessarily 1.)

Supertype Hierarchy

_Orthogonality <: _KOrthogonality <: Any

Notes

Since this type is semantically unique inasmuch as it simply encodes the information k = 1, we define the singleton instance _ORTHOGONALITY to always be used when needed in favor of instantiating an identical object every time.

SDiagonalizability._QOBasisSearchNodeData — Type

mutable struct _QOBasisSearchNodeData

TODO: Write here

SDiagonalizability._QuasiOrthogonality — Type

struct _QuasiOrthogonality

The property of quasi-orthogonality for a collection of vectors.

Recall that an (ordered) collection of vectors v₁, v₂, ..., vₙ is said to be quasi-orthogonal if we have the inner product ⟨vᵢ, vⱼ⟩ = 0 whenever |i - j| ≥ 2 (i.e., if every pair of vectors at least 2 indices apart is orthogonal). This is equivalent to the vectors' Gram matrix being tridiagonal [JP25, p. 313].

Properties

k::Int: the k-orthogonality parameter. (Necessarily 2.)

Supertype Hierarchy

_QuasiOrthogonality <: _KOrthogonality <: Any

Notes

Since this type is semantically unique inasmuch as it simply encodes the information k = 2, we define the singleton instance _QUASI_ORTHOGONALITY to always be used when needed in favor of instantiating an identical object every time.

SDiagonalizability._TypedLaplacian — Type

abstract type _TypedLaplacian

An abstract type representing a classified Laplacian matrix of an undirected graph.

Interface

Concrete subtypes of _TypedLaplacian must implement the following fields:

matrix::Matrix{Int}: the Laplacian matrix of the graph.

Subtypes

_NullGraphLaplacian`: a Laplacian wrapper for the (unique) graph with no nodes.

_EmptyGraphLaplacian: a Laplacian wrapper for any graph with no edges on n ≥ 1 nodes.

_CompleteGraphLaplacian: a Laplacian wrapper for any graph on n ≥ 2 nodes where every pair of nodes is connected by an edge and all edges possess the same weight.

_ArbitraryGraphLaplacian: a Laplacian wrapper for any graph on n ≥ 3 nodes with no particular strcuture of relevance in the context of investigating {-1,0,1}-spectra.

Notes

Given that this type and all its subtypes are privately encapsulated, we expect _TypedLaplacian instances to be created exclusively via the _cast_to_typed_laplacian factory function, which copies each input matrix to avoid shared mutability and converts it to a Matrix{Int} in the process. Therefore, it is safe to restrict the type of matrix field (in each concrete subtype) to Matrix{Int} in lieu of the more generic AbstractMatrix{<:Integer}. This decision may contribute marginal performance benefits via reduced type complexity and is highly unlikely to cause any issues (as this type is never to be exported to the public API). It also serves as an additional safeguard against directly passing a (wrongly typed) matrix to a subtype's constructor rather than calling _cast_to_typed_laplacian.

On a related note, it seems prudent to discuss the intended use case of _TypedLaplacian. The interface and its subtypes are meant to support _cast_to_typed_laplacian. This factory, in turn, determines which algorithm laplacian_spectra_01neg is to implement when computing data on the Laplacian matrix of a given graph, as different types of graphs exhibit different spectral properties which we may exploit to improve performance.

SDiagonalizability._WeakOrthogonality — Type

struct _WeakOrthogonality

The property of "weak orthogonality" for a collection of vectors.

In particular, "weak orthogonality" is an ad hoc term used to refer to the property of k-orthogonality for k > 2. Recall that an (ordered) collection of vectors v₁, v₂, ..., vₙ is said to be k-orthogonal if we have the inner product ⟨vᵢ, vⱼ⟩ = 0 whenever |i - j| ≥ k (i.e., if every pair of vectors at least k indices apart is orthogonal). This is equivalent to the vectors' Gram matrix having bandwidth at most k.

Fields

k::Int: the k-orthogonality parameter. (Necessarily a positive integer greater than 2.)

Supertype Hierarchy

_WeakOrthogonality <: _KOrthogonality <: Any

Constructors

_WeakOrthogonality(::Integer): constructs a _WeakOrthogonality object with the given k-orthogonality parameter, casting it to an Int for type stability. Throws a DomainError if the argument is less than or equal to 2.

Notes

The term "weak orthogonality" is not standard terminology in the literature, but it is used here to emphasize the weaker nature of this property compared to orthogonality and quasi-orthogonality. It is an ad hoc term coined for this module and is not intended to be formally introduced in the broader literature. See also the documentation for parent type _KOrthogonality for further discussion of this topic, as well as justification for enforcing the k field to be an Int rather than the more generic Integer.

SDiagonalizability._assert_matrix_is_undirected_laplacian — Method

_assert_matrix_is_undirected_laplacian(L)

Validate that L is the Laplacian matrix of an undirected, possibly weighted graph.

Arguments

L::AbstractMatrix{<:Integer}: the purported Laplacian matrix.

Returns

nothing: if the check is passed, no output is produced.

Throws

DomainError: if L is not symmetric or has nonzero row sums.

Examples

The Laplacian matrix of the (undirected) star graph on 5 vertices passes the check:

julia> L = [ 4  -1  -1  -1  -1;
            -1   1   0   0   0;
            -1   0   1   0   0;
            -1   0   0   1   0;
            -1   0   0   0   1]
5×5 Matrix{Int64}:
  4  -1  -1  -1  -1
 -1   1   0   0   0
 -1   0   1   0   0
 -1   0   0   1   0
 -1   0   0   0   1

julia> isnothing(SDiagonalizability._assert_matrix_is_undirected_laplacian(L))
true

The adjacency matrix of the (undirected) cycle graph on 4 vertices is symmetric but has nonzero row sums, so it fails the check:

julia> A = [0  1  0  1;
            1  0  1  0;
            0  1  0  1;
            1  0  1  0]
4×4 Matrix{Int64}:
 0  1  0  1
 1  0  1  0
 0  1  0  1
 1  0  1  0

julia> SDiagonalizability._assert_matrix_is_undirected_laplacian(A)
ERROR: DomainError with [0 1 0 1; 1 0 1 0; 0 1 0 1; 1 0 1 0]:
Matrix has nonzero row sums; cannot be an (undirected) Laplacian
[...]

Both the in-degree and out-degree Laplacian matrices of this random tournament digraph have zero row sums but are not symmetric, so they fail the check. (These are the two standard ways of extending the concept of the Laplacian to directed graphs [VL20, p. 196].)

julia> G = random_tournament_digraph(3; seed=87)
{3, 3} directed simple Int64 graph

julia> L_in = laplacian_matrix(G; dir=:in)
3×3 SparseArrays.SparseMatrixCSC{Int64, Int64} with 5 stored entries:
  ⋅  ⋅   ⋅
 -1  2  -1
 -1  ⋅   1

julia> SDiagonalizability._assert_matrix_is_undirected_laplacian(L_in)
ERROR: DomainError with sparse([2, 3, 2, 2, 3], [1, 1, 2, 3, 3], [-1, -1, 2, -1, 1], 3, 3):
Matrix is not symmetric; cannot be an (undirected) Laplacian
[...]

julia> L_out = laplacian_matrix(G; dir=:out)
3×3 SparseArrays.SparseMatrixCSC{Int64, Int64} with 5 stored entries:
 2  -1  -1
 ⋅   ⋅   ⋅
 ⋅  -1   1

julia> SDiagonalizability._assert_matrix_is_undirected_laplacian(L_out)
ERROR: DomainError with sparse([1, 1, 3, 1, 3], [1, 2, 2, 3, 3], [2, -1, -1, -1, 1], 3, 3):
Matrix is not symmetric; cannot be an (undirected) Laplacian
[...]

Notes

If edges are to be bidirectional, then L must be symmetric. L must also have zero row sums, since the (i, i)-th entry is the weighted degree of node i (the sum of all incident edges' weights) and the (i, j)-th entry for i ≠ j is the negation of the weight of edge (i, j) (or simply 0, if no such edge exists).

Given the highly optimized, lazy, zero-allocation implementation of LinearAlgebra.issymmetric, the symmetry check is performed first. (Both steps are O(n²) in the worst case, but testing for symmetry is far more performant in practice.) This also allows us to (also lazily) check for nonzero column sums rather than nonzero row sums (since these are equivalent for symmetric matrices) in the second step, taking advantage of Julia's column-major storage model.

At first blush, it may seem as though the choice of DomainError over something like ArgumentError (or even simply the return of a boolean) constitutes poor design. However, this is informed by the simple ad hoc use of this function to validate inputs for other functions requiring Laplacian matrices. Certainly, this function is never meant to be publicly exposed on its own.

SDiagonalizability._cast_to_typed_laplacian — Method

_cast_to_typed_laplacian(L)

Classify the Laplacian matrix L and wrap it in a _TypedLaplacian object.

It is first verified that L is indeed a Laplacian matrix by _assert_matrix_is_undirected_laplacian, which throws a DomainError otherwise. It is then classified based on any properties which may be exploited in computing data on its {-1,0,1}-spectrum.

Arguments

L::AbstractMatrix{<:Integer}: the Laplacian matrix to classify.

Returns

TL::_TypedLaplacian: the Laplacian wrapped in a concrete _TypedLaplacian subtype associated with the category of the graph represented by L. In the case of complete graphs, TL also contains data on the (necessarily uniform) weight of each edge.

Examples

Correctly recognizes the Laplacian matrix of the null graph:

julia> G = Graph(0)
{0, 0} undirected simple Int64 graph

julia> L = laplacian_matrix(G)
0×0 SparseArrays.SparseMatrixCSC{Int64, Int64} with 0 stored entries

julia> TL = SDiagonalizability._cast_to_typed_laplacian(L)
SDiagonalizability._NullGraphLaplacian(Matrix{Int64}(undef, 0, 0))

julia> isa(TL, SDiagonalizability._NullGraphLaplacian)
true

Correctly recognizes the Laplacian matrix of the empty graph on 3 nodes:

julia> G = Graph(3)
{3, 0} undirected simple Int64 graph

julia> L = laplacian_matrix(G)
3×3 SparseArrays.SparseMatrixCSC{Int64, Int64} with 0 stored entries:
 ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅

julia> TL = SDiagonalizability._cast_to_typed_laplacian(L)
SDiagonalizability._EmptyGraphLaplacian([0 0 0; 0 0 0; 0 0 0])

julia> isa(TL, SDiagonalizability._EmptyGraphLaplacian)
true

Correctly recognizes the Laplacian matrix of the complete graph on 4 nodes:

julia> G = complete_graph(4)
{4, 6} undirected simple Int64 graph

julia> L = laplacian_matrix(G)
4×4 SparseArrays.SparseMatrixCSC{Int64, Int64} with 16 stored entries:
  3  -1  -1  -1
 -1   3  -1  -1
 -1  -1   3  -1
 -1  -1  -1   3

julia> TL = SDiagonalizability._cast_to_typed_laplacian(L)
SDiagonalizability._CompleteGraphLaplacian([3 -1 -1 -1; -1 3 -1 -1; -1 -1 3 -1; -1 -1 -1 3], -1)

julia> isa(TL, SDiagonalizability._CompleteGraphLaplacian)
true

Correctly recognizes the Laplacian matrix of this random generic graph:

julia> G = erdos_renyi(5, 8; seed=87)
{5, 8} undirected simple Int64 graph

julia> L = laplacian_matrix(G)
5×5 SparseArrays.SparseMatrixCSC{Int64, Int64} with 21 stored entries:
  3   ⋅  -1  -1  -1
  ⋅   2  -1  -1   ⋅
 -1  -1   4  -1  -1
 -1  -1  -1   4  -1
 -1   ⋅  -1  -1   3

julia> TL = SDiagonalizability._cast_to_typed_laplacian(L)
SDiagonalizability._ArbitraryGraphLaplacian([3 0 … -1 -1; 0 2 … -1 0; … ; -1 -1 … 4 -1; -1 0 … -1 3])

julia> isa(TL, SDiagonalizability._ArbitraryGraphLaplacian)
true

Notes

Right now, the possible types of Laplacian matrices (or, to be more precise, the graphs they represent) are:

_NullGraphLaplacian: the (unique) graph with no nodes.

_EmptyGraphLaplacian: any graph with no edges on n ≥ 1 nodes.

_CompleteGraphLaplacian: any graph on n ≥ 2 nodes where every pair of nodes is connected by an edge and all edges possess the same weight.

_ArbitraryGraphLaplacian: any graph on n ≥ 3 nodes with no particular strcuture of relevance in the context of investigating {-1,0,1}-spectra.

SDiagonalizability._classify_orthogonality_property — Method

_classify_orthogonality_property(k)

Classifies the k-orthogonality property based on the given k parameter.

When searching for a k-orthogonal S-basis of a given Laplacian eigenspace, the family of values to which our k parameter belongs informs our choice of algorithm.

Arguments

k::Integer: the k-orthogonality parameter to classify. Must be a positive integer.

Returns

::_KOrthogonality: An instance of a concrete _KOrthogonality subtype corresponding to k.

Throws

DomainError: if k is not a positive integer.

Examples

julia> SDiagonalizability._classify_orthogonality_property(1)
SDiagonalizability._Orthogonality()

julia> SDiagonalizability._classify_orthogonality_property(2)
SDiagonalizability._QuasiOrthogonality()

julia> SDiagonalizability._classify_orthogonality_property(3)
SDiagonalizability._WeakOrthogonality(3)

julia> SDiagonalizability._classify_orthogonality_property(13)
SDiagonalizability._WeakOrthogonality(13)

Notes

See the _find_k_orthogonal_basis documentation for more details on how this function is used in the context of finding a k-orthogonal basis.

SDiagonalizability._extract_independent_cols — Method

_extract_independent_cols(A)

Return a (not necessarily unique) independent spanning subset of the columns of A.

Computing a rank-revealing (pivoted) QR decomposition of A, the scaling coefficients from the orthogonalization process are used to determine the rank (rather than recompute it with an SVD), while the pivots are used to extract a spanning set of independent columns.

The rank-revealing Businger–Golub QR algorithm is used for the pivoting strategy, appending the "most independent" column with respect to the current set of pivots at each step via Householder transformations [BG65, pp. 269–70].

Arguments

A::AbstractMatrix{T<:Integer}: the matrix whose independent columns to extract.

Returns

::AbstractMatrix{T}: a spanning set of independent columns of A.

Examples

Observe how columns with greater Euclidean norms are given priority in the pivot ordering:

julia> A = [3  0  0  0  2  1   5   0
            0  3  0  0  2  1  -5   0
            0  0  3  0  2  1   5   4
            0  0  0  3  2  1   0  -4
            0  0  0  0  0  0   0   0]
5×8 Matrix{Int64}:
 3  0  0  0  2  1   5   0
 0  3  0  0  2  1  -5   0
 0  0  3  0  2  1   5   4
 0  0  0  3  2  1   0  -4
 0  0  0  0  0  0   0   0

julia> SDiagonalizability._extract_independent_cols(A)
5×4 Matrix{Int64}:
  5   0  2  3
 -5   0  2  0
  5   4  2  0
  0  -4  2  0
  0   0  0  0

Notes

Since we already need a pivoted QR decomposition to identify independent columns of A (or, rather, to order the columns in such a way that the first rank(A) ones are guaranteed to be independent), it makes sense to use data from the resulting factorization object to compute the rank of A rather than compute a separate SVD. We thus count the nonzero scaling coefficients—that is, the diagonal entries of the R matrix in A = QR—to determine the rank, similarly to how we count the nonzero singular values in an SVD.

It is worth noting that we manually specify a higher relative tolerance for this rank computation. Further discussion can be found in the _rank_rtol documentation, but in short, a critical part of the formula for LinearAlgebra.rank's default rtol uses the minimum dimension of the input matrix. This may result in rank overestimation for tall-and-skinny and short-and-fat matrices (precisely the type we expect to encounter when dealing with all {-1,0,1}-eigenvectors of a Laplacian matrix, which is the intended use case of this helper function in this package). Our replacement tolerance, on the other hand, is a widely accepted standard in numerical analysis which uses the maximum dimension instead [PTVF07, p. 795].

SDiagonalizability._find_basis_indices — Method

_find_basis_indices(
    curr_indices,
    column_space,
    column_rank,
    G::Graph,
    H::Graph,
    prop::_WeakOrthogonality,
)

TODO: Write here

SDiagonalizability._find_basis_indices — Method

_find_basis_indices(curr_indices, column_space, column_rank, prop::_Orthogonality)

TODO: Write here

SDiagonalizability._find_basis_indices — Method

_find_basis_indices(
    curr_indices,
    column_space,
    column_rank,
    nodes,
    union_find,
    prop::_QuasiOrthogonality,
)

TODO: Write here

SDiagonalizability._find_basis_with_property — Method

_find_basis_with_property(column_space, column_rank, prop::_KOrthogonality)

TODO: Write here

SDiagonalizability._find_basis_with_property — Method

_find_basis_with_property(column_space, column_rank, prop:::_Orthogonality)

TODO: Write here, and add comments

SDiagonalizability._find_basis_with_property — Method

_find_basis_with_property(column_space, column_rank, prop::_QuasiOrthogonality)

TODO: Write here, and add comments

SDiagonalizability._find_basis_with_property — Method

_find_basis_with_property(column_space, column_rank, prop::_WeakOrthogonality)

TODO: Write here

SDiagonalizability._find_indices_1neg — Method

_find_indices_1neg(eigvecs_01neg)

Find the indices of {-1,1}-eigenvectors given a map of eigenspaces.

More precisely, given a map from each eigenvalue to all {-1,0,1}-vectors in the associated eigenspace, this function returns a map from each eigenvalue to the indices of the {-1,1}-vectors in each value from the input map.

It is, of course, assumed that the input map indeed contains only {-1,0,1}-matrices as values, and that it contains 0 as a key.

Arguments

eigvecs_01neg::AbstractDict{Int,<:AbstractMatrix{Int}}: a map from each eigenvalue to all {-1,0,1}-eigenvectors in the associated eigenspace.

Returns

::AbstractDict{Int,Vector{Int}}: a map from each eigenvalue eigval to the indices of the {-1,1}-columns in eigvecs_01neg[eigval].

Examples

julia> eigvecs_01neg = Dict(
           0 => [1  1   1;
                 1  1   1;
                 1  1   1;
                 1  0  -1],
           2 => [ 1   0;
                  0   1;
                 -1  -1;
                  0   0],
       )
Dict{Int64, Matrix{Int64}} with 2 entries:
  0 => [1 1 1; 1 1 1; 1 1 1; 1 0 -1]
  2 => [1 0; 0 1; -1 -1; 0 0]

julia> SDiagonalizability._find_indices_1neg(eigvecs_01neg)
Dict{Int64, Vector{Int64}} with 2 entries:
  0 => [1, 3]
  2 => []

Notes

It is important to note that this function implicitly assumes that the input map already constitutes valid {-1,0,1}-spectra for some undirected Laplacian matrix. For instance, the assumption that all eigenvectors corresponding to nonzero eigenvalues are orthogonal to the all-ones vector (which must be in the kernel) is integral to the correctness of this algorithm. Since _find_indices_1neg is ultimately but a helper function for laplacian_spectra_01neg, which already validates that the input passed is valid, we choose not to double-check it here for performance reasons. (Usually, it would be good practice to do so, but the potential sheer size of eigvecs_01neg, combined with the time-sensitive nature of the algorithm, precludes the possibility in this particular case.)

Another design choice worthy of discussion is to encode (per eigenspace) the data on which columns have entries exclusively from {-1,1} as a Vector{Int} rather than a BitVector. Vector{Int}'s become more memory efficient than BitVector's to store indices when the density of trues is less than 1 / 64, which is far greater than the proportion of {-1,1}-columns in most (albeit not all) cases, simply based on our numerical results.

SDiagonalizability._find_k_orthogonal_basis — Method

_find_k_orthogonal_basis(column_space, column_rank, k)

Find a k-orthogonal spanning subset of the column set of a matrix.

This function does not allow for any k-orthogonal basis of column_space (in which case it would be trivial to construct a pairwise orthogonal basis via a QR decomposition). Instead, it restricts its search to spanning subsets comprised exclusively of columns already in column_space, returning nothing if no such basis exists.

Arguments

column_space::T<:AbstractMatrix{Int}: the matrix whose column space is search for a k-orthogonal basis.

column_rank::Int: the rank of the column space of column_space, pre-computed for efficiency.

k::Integer: the minimum k-orthogonality parameter of the desired basis. (Must be a positive integer.)

Returns

::Union{Nothing,T}: a k-orthogonal spanning subset of the columns of column_space, with the same type as column_space. If no such basis exists, this is simply nothing.

Notes

TODO: Write here. In particular, comment on the "basis" naming and perhaps k. Also, comment on typing decisions. (Should k just be Int?)

SDiagonalizability._pot_kernel_eigvecs_01neg — Method

_pot_kernel_eigvecs_01neg(n)

Lazily generate all potential kernel {-1,0,1}-eigenvectors of an order n Laplacian.

Each vector is normalized so that its first nonzero entry is 1, enforcing pairwise linear independence between all generated vectors.

Arguments

n::Integer: the order of the Laplacian matrix of some undirected graph for which to find potential kernel eigenvectors.

Returns

eigvec_generator::Iterators.Flatten{<:Base.Generator}: a lazily evaluated iterator over all {-1,0,1}-vectors in ℝⁿ, unique up to span.

Throws

DomainError: if n is negative.

Examples

Generate all potential kernel eigenvectors for an order 3 Laplacian matrix:

julia> hcat(SDiagonalizability._pot_kernel_eigvecs_01neg(3)...)
3×13 Matrix{Int64}:
  1   1   1   1  1  1   1  1  1   0  0  0  0
 -1  -1  -1   0  0  0   1  1  1   1  1  1  0
 -1   0   1  -1  0  1  -1  0  1  -1  0  1  1

Notes

The number of potential kernel eigenvectors (unique up to span) for an order n Laplacian matrix is given by (3ⁿ - 1) / 2. See also the relevant OEIS sequence [Slo25].

Regrettably, the implementation here is rather clunky and unidiomatic, but it is worth noting that eigenvector generation (and subsequent processing by laplacian_spectra_01neg) is one of two major bottlenecks in the overall S-bandwidth minimization algorithm. Given how much potential there is for optimization in this piece of code, we thus prioritize performance over readability in this particular case, making every effort to include inline comments wherever clarification may be needed.

SDiagonalizability._pot_nonkernel_eigvecs_01neg — Method

_pot_nonkernel_eigvecs_01neg(n)

Lazily generate all potential non-kernel {-1,0,1}-eigenvectors of an order n Laplacian.

Each vector is normalized so that its first nonzero entry is 1, enforcing pairwise linear independence between all generated vectors. Since all Laplacian matrices have pairwise orthogonal eigenspaces and the all-ones vector is always in the kernel, every non-kernel {-1,0,1}-eigenvector must also have an equal number of -1's and 1's.

Arguments

n::Integer: the order of the Laplacian matrix of some undirected graph for which to find potential non-kernel eigenvectors.

Returns

eigvec_generator::Iterators.Flatten{<:Base.Generator}: a lazily evaluated iterator over all {-1,0,1}-vectors in ℝⁿ orthogonal to the all-ones kernel vector, unique up to span.

Throws

DomainError: if n is negative.

Examples

Generate all potential non-kernel eigenvectors of an order 4 Laplacian matrix:

julia> hcat(SDiagonalizability._pot_nonkernel_eigvecs_01neg(4)...)
4×9 Matrix{Int64}:
  1   1   1   1   1   1   0   0   0
 -1   0   0  -1  -1   1   1   1   0
  0  -1   0  -1   1  -1  -1   0   1
  0   0  -1   1  -1  -1   0  -1  -1

Notes

The number of potential non-kernel eigenvectors (unique up to span) for an order n Laplacian matrix is, by non-trivial combinatorial arguments, equal to the number of humps in all Motzkin paths of length n. See also the relevant OEIS sequence [Deu25].

Regrettably, the implementation here is rather clunky and unidiomatic, but it is worth noting that eigenvector generation (and subsequent processing by laplacian_spectra_01neg) is one of two major bottlenecks in the overall S-bandwidth minimization algorithm. Given how much potential there is for optimization in this piece of code, we thus prioritize performance over readability in this particular case, making every effort to include inline comments wherever clarification may be needed.

SDiagonalizability._rank_rtol — Method

function _rank_rtol(A::AbstractMatrix{<:Real})

Return a reasonable relative tolerance for computing matrix rank via SVD or QRD.

The output is intended to be passed as a keyword argument to LinearAlgebra.rank. LinearAlgebra.eigtype is used to determine the appropriate machine epsilon for the element type of A when eltype(A) is not an AbstractFloat.

Arguments

A::AbstractMatrix{<:Real}: the matrix for which to compute a tolerance.

Returns

tol::Float64: a reasonable relative tolerance for computing matrix rank via SVD or QRD. This scales proportionally to the maximum dimension of A.

Notes

LinearAlgebra.rank's default rtol of min(m,n) * ϵ for computing the rank of an m × n matrix may result in overestimating rank when |m - n| ≫ 0, since condition number (which determines how numerically stable SVD and QRD are) grows with both dimensions [CD05, p. 603]. Given that we often deal with short-and-fat matrices in this package (particularly when processing all {-1,0,1}-eigenvectors of a Laplacian matrix), we turn instead to the same relative tolerance used by NumPy's and MATLAB's rank functions—max(m,n) * ϵ [Num25, MAT25]. (Indeed, this is a widely adopted standard across the field of numerical analysis [PTVF07, p. 795].)

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_ArbitraryGraphLaplacian)

TODO: Write here

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_CompleteGraphLaplacian)

TODO: Write here

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_EmptyGraphLaplacian)

TODO: Write here

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_NullGraphLaplacian)

TODO: Write here

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_TypedLaplacian)

TODO: Write here

SDiagonalizability.rank — Function

rank(A::QRPivoted{<:Any, T}; atol::Real=0, rtol::Real=min(n,m)*ϵ) where {T}

Compute the numerical rank of the QR factorization A by counting how many diagonal entries of A.factors are greater than max(atol, rtol*Δ₁) where Δ₁ is the largest calculated such entry. This is similar to the LinearAlgebra.rank(::AbstractMatrix) method insofar as it counts the number of (numerically) nonzero coefficients from a matrix factorization, although the default method uses an SVD instead of a QR factorization. Like LinearAlgebra.rank(::SVD), this method also re-uses an existing matrix factorization.

Computing rank via QR factorization should almost always produce the same results as via SVD, although this method may be more prone to overestimating the rank in pathological cases where the matrix is ill-conditioned. It is also worth noting that it is generally faster to compute a QR factorization than an SVD, so this method may be preferred when performance is a concern.

atol and rtol are the absolute and relative tolerances, respectively. The default relative tolerance is n*ϵ, where n is the size of the smallest dimension of A and ϵ is the eps of the element type of A.

References

[BG65]

P. Businger and G. H. Golub. Linear Least Squares Solutions by Householder Transformations. Numerische Mathematik 7, 269–276 (1965).

[CD05]

Z. Chen and J. Dongarra. Condition Numbers of Gaussian Random Matrices. SIAM Journal on Matrix Analysis and Applications 27, 603–620 (2005).

[Deu25]

E. Deutsch. Number of humps in all Motzkin paths of length n. Entry A097861 in The On-Line Encyclopedia of Integer Sequences (2025). Accessed: 2025-05-22.

[Maf14]

L. O. Mafteiu-Scai. The Bandwidths of a Matrix. A Survey of Algorithms. Annals of West University of Timisoara - Mathematics and Computer Science 52, 183–223 (2014).

[PTVF07]

W. H. Press, S. A. Teukolsky, W. T. Vetterling and B. P. Flannery. Numerical Recipes: The Art of Scientific Computing. 3rd Edition (Cambridge University Press, Cambridge, UK, 2007). ISBN: 978-0-521-88068-8.

[Slo25]

N. J. Sloane, a(n) = (3^n - 1)/2. Entry A003462 in The On-Line Encyclopedia of Integer Sequences (2025). Accessed: 2025-05-22.

[VL20]

J. J. Veerman and R. Lyons. A Primer on Laplacian Dynamics in Directed Graphs. Nonlinear Phenomena in Complex Systems 23, 196–206 (2020).

[MAT25]

MATLAB Developers, rank. MATLAB reference documentation – R2025a (2025). Accessed: 2025-05-29.

[Num25]

NumPy Developers, numpy.linalg.matrix_rank. NumPy reference documentation – v2.2 (2025). Accessed: 2025-05-22.