SDiagonalizability.jl – Public API

Documentation for SDiagonalizability's public API.

Note

The following documentation covers only the public API of the package. For internal details, see the private API documentation.

SDiagonalizability.SDiagonalizability — Module

SDiagonalizability

A dynamic algorithm to test a graph for S-diagonalizability and minimize its S-bandwidth.

SDiagonalizability.jl packages the first non-naïve deterministic algorithm to minimize the S-bandwidth of a quantum network (or, to be more precise, its graph representation), written in Julia. Given some finite S ∈ Zⁿ, the S-bandwidth of an (undirected) graph G with Laplacian L(G) ∈ Rⁿˣⁿ is the minimum integer k ≥ 1 such that L(G) = PDP⁻¹ for some diagonal D ∈ Rⁿˣⁿ and P ∈ Sⁿˣⁿ with [PᵀP]ᵢⱼ = 0 whenever |i - j| ≥ k. For specific choices of S (namely S = {-1,1} and S = {-1,0,1}), a quantum network's S-bandwidth has been shown to be an indicator of high state transfer fidelity due to automorphic properties of its graph representation.

Full documentation is available for the latest development version of this package.

SDiagonalizability.LaplacianSpectrum01Neg — Type

struct LaplacianSpectrum01Neg{R,S<:R,T<:R,U<:R,V<:R}

Data on the {-1,0,1}-spectrum of some Laplacian matrix.

Each instance encapsulates the Laplacian itself, whether the Laplacian is {-1,0,1}- and {-1,1}-diagonalizable, each eigenvalue's multiplicity, and the associated {-1,0,1}- and {-1,1}-eigenvectors, and eigenbases. If the Laplacian is not spectrum integral, then it cannot have any {-1,0,1}-eigenvectors outside of the kernel [JP25, p. 312], so much of this data is simply nothing.

Type Parameters

R<:Union{Nothing,OrderedDict{Int}}: the "common upper bound" for all five map-like fields, ensuring that they are either all of the same supertype. Either Nothing (if and only if the Laplacian is not spectrum integral, since we cannot map non-integer eigenvalues to data) or OrderedDict{Int} (if all eigenvalues are indeed integers).

S<:R: the type of the multiplicities field. Either Nothing or an OrderedDict{Int,Int}.

T<:R: the type of the eigvecs_01neg field. Either Nothing or an OrderedDict{Int,AbstractMatrix{Int}}.

U<:R: the type of the indices_1neg field. Either Nothing or an OrderedDict{Int,Vector{Int}}.

V<:R: the type of the bases_01neg and bases_1neg fields. Either Nothing or an OrderedDict{Int,Union{Nothing,Matrix{Int}}}.

Fields

laplacian_matrix::Matrix{Int}: the Laplacian matrix.

diagonalizable_01neg::Bool: whether the Laplacian is {-1,0,1}-diagonalizable.

diagonalizable_1neg::Bool: whether the Laplacian is {-1,1}-diagonalizable.

multiplicities::S: a map from each eigenvalue to its multiplicity, sorted first by ascending multiplicity then by ascending eigenvalue. (If the Laplacian is not spectrum integral, this field is nothing.)

eigvecs_01neg::T: a map from each eigenvalue to its {-1,0,1}-eigenvectors, stored as columns of a matrix. (If the Laplacian is not spectrum integral, this field is nothing.)

indices_1neg::U: a map from each eigenvalue to the indices of the {-1,1}-columns in eigvecs_01neg[eigval]. (If the Laplacian is not spectrum integral, this field is nothing.)

bases_01neg::V: a map from each eigenvalue to a {-1,0,1}-basis for the corresponding eigenspace if one exists, or to nothing otherwise. (If the Laplacian is not spectrum integral, this field is nothing.)

bases_1neg::V: a map from each eigenvalue to a {-1,1}-basis for the corresponding eigenspace if one exists, or to nothing otherwise. (If the Laplacian is not spectrum integral, this field is nothing.)

Constructors

LaplacianSpectrum01Neg(::Matrix{Int}, ::Bool, ::Bool, ::Union{Nothing,OrderedDict{Int,_Eigenspace01Neg}}): constructs a LaplacianSpectrum01Neg instance from the given Laplacian matrix and associated data on its {-1,0,1}-eigenspaces. When the last argument is not nothing (indicating spectrum integrality), this constructors flattens the data stored in the _Eigenspace01Neg instances and exposes it via top-level fields. (See the _Eigenspace01Neg documentation for more details.)

Notes

TODO: Discuss potential use cases for this beyond the s_bandwidth algorithm

SDiagonalizability.SpectrumIntegralResult — Type

struct SpectrumIntegralResult

Data on whether a matrix is spectrum integral (i.e., whether its eigenvalues are integers).

This struct also contains a map from each eigenvalue to its multiplicity, provided that the eigenvalues are indeed all integers. (Otherwise, the associated field is simply nothing.)

Fields

matrix::Matrix{Int}: the matrix whose eigenvalues and their integrality are of interest.

spectrum_integral::Bool: whether the eigenvalues of matrix are all integers.

multiplicities::Union{Nothing,OrderedDict{Int,Int}}: a map from each eigenvalue to its multiplicity, sorted first by ascending multiplicity then by ascending eigenvalue. (This field is nothing if spectrum_integral is false.)

Notes

If an undirected graph with integer edge weights is {-1,0,1}-diagonalizable (or, more restrictively, {-1,1}-diagonalizable), then its Laplacian matrix has integer eigenvalues [JP25, p. 312]. Hence, validating Laplacian integrality serves as a useful screening step in this package's principal S-bandwidth minimization algorithm.

SDiagonalizability.check_spectrum_integrality — Method

check_spectrum_integrality(A)

Check whether the eigenvalues of A are integers (up to floating-point error).

If the eigenvalues are indeed all integers, then an eigenvalue-multiplicity map is constructed as well.

Arguments

A::AbstractMatrix{<:Integer}: the matrix whose eigenvalues to check for integrality.

Returns

::SpectrumIntegralResult: a struct containing the following fields:

matrix: the A matrix, copied to avoid shared mutability.

spectrum_integral: whether the eigenvalues of A are integers.

multiplicities: a map from each eigenvalue to its multiplicity, sorted first by ascending multiplicity then by ascending eigenvalue. (This field is nothing if the eigenvalues are not all integers.)

Examples

Confirm that the rotation matrix by π/2 radians counterclockwise is not spectrum integral (rather, it has eigenvalues ±i [Joy15, p. 1]):

julia> R = Int8.([0 -1; 1 0])
2×2 Matrix{Int8}:
 0  -1
 1   0

julia> res = check_spectrum_integrality(R);

julia> res.matrix
2×2 Matrix{Int64}:
 0  -1
 1   0

julia> res.spectrum_integral
false

julia> isnothing(res.multiplicities)
true

Confirm that the adjacency matrix of the Petersen graph is spectrum integral, with correct eigenvalues and multiplicities of {3: 1, -2: 4, 1: 5} [Fox09, p. 2]:

julia> G = smallgraph(:petersen)
{10, 15} undirected simple Int64 graph

julia> A = adjacency_matrix(G)
10×10 SparseArrays.SparseMatrixCSC{Int64, Int64} with 30 stored entries:
 ⋅  1  ⋅  ⋅  1  1  ⋅  ⋅  ⋅  ⋅
 1  ⋅  1  ⋅  ⋅  ⋅  1  ⋅  ⋅  ⋅
 ⋅  1  ⋅  1  ⋅  ⋅  ⋅  1  ⋅  ⋅
 ⋅  ⋅  1  ⋅  1  ⋅  ⋅  ⋅  1  ⋅
 1  ⋅  ⋅  1  ⋅  ⋅  ⋅  ⋅  ⋅  1
 1  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  1  1  ⋅
 ⋅  1  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  1  1
 ⋅  ⋅  1  ⋅  ⋅  1  ⋅  ⋅  ⋅  1
 ⋅  ⋅  ⋅  1  ⋅  1  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  ⋅  1  ⋅  1  1  ⋅  ⋅

julia> res = check_spectrum_integrality(A);

julia> res.matrix
10×10 Matrix{Int64}:
 0  1  0  0  1  1  0  0  0  0
 1  0  1  0  0  0  1  0  0  0
 0  1  0  1  0  0  0  1  0  0
 0  0  1  0  1  0  0  0  1  0
 1  0  0  1  0  0  0  0  0  1
 1  0  0  0  0  0  0  1  1  0
 0  1  0  0  0  0  0  0  1  1
 0  0  1  0  0  1  0  0  0  1
 0  0  0  1  0  1  1  0  0  0
 0  0  0  0  1  0  1  1  0  0

julia> res.spectrum_integral
true

julia> res.multiplicities
OrderedCollections.OrderedDict{Int64, Int64} with 3 entries:
  3  => 1
  -2 => 4
  1  => 5

Notes

If an undirected graph with integer edge weights is {-1,0,1}-diagonalizable (or, more restrictively, {-1,1}-diagonalizable), then its Laplacian matrix has integer eigenvalues [JP25, p. 312]. Hence, validating Laplacian integrality serves as a useful screening step in this package's principal S-bandwidth minimization algorithm.

It is, perhaps, an odd choice to sort the eigenvalue/multiplicity pairs in this file rather than in s_bandwidth—after all, in the context of the overarching S-bandwidth algorithm, this ordering is only ever used to determine which eigenspaces are searched for S-bases first. However, the inclusion of check_spectrum_integrality in the public API motivates a consistent, natural ordering of the multiplicites map in each (potentially user-facing) SpectrumIntegralResult instance.

Of course, we make it a point to still sort multiplicities as desired (first by ascending multiplicity then by ascending eigenvalue) in s_bandwidth itself; this seems more robust than relying on check_spectrum_integrality to do so or even folding the logic into the SpectrumIntegralResult inner constructor.

SDiagonalizability.laplacian_spectra_01neg — Method

laplacian_spectra_01neg(L::AbstractMatrix{<:Integer})

Compute data on the {-1,0,1}- and {-1,1}-spectrum of some Laplacian matrix L.

The original matrix, whether it is {-1,0,1}- and {-1,1}-diagonalizable, and further data on the eigenvalues and their corresponding {-1,0,1}- and {-1,1}-eigenvectors are all stored in a LaplacianSpectrum01Neg instance and returned.

Arguments

L::AbstractMatrix{<:Integer}: the Laplacian matrix in whose {-1,0,1}- and {-1,1}-spectrum we are interested.

Returns

::LaplacianSpectrum01Neg: a struct containing the following fields:

laplacian_matrix: the L matrix, copied to avoid shared mutability.

diagonalizable_01neg: whether the Laplacian is {-1,0,1}-diagonalizable.

diagonalizable_1neg: whether the Laplacian is {-1,1}-diagonalizable.

multiplicities: a map from each eigenvalue to its multiplicity, sorted first by ascending multiplicity then by ascending eigenvalue. (This field is nothing if the Laplacian is not spectrum integral.)

eigvecs_01neg: a map from each eigenvalue to its {-1,0,1}-eigenvectors. (This field is nothing if the Laplacian is not spectrum integral.)

indices_1neg: a map from each eigenvalue to the indices of the {-1,1}-columns in eigvecs_01neg[eigval]. (This field is nothing if the Laplacian is not spectrum integral.)

bases_01neg: a map from each eigenvalue to a {-1,0,1}-basis for the corresponding eigenspace. (This field is nothing if the Laplacian is not spectrum integral.)

bases_1neg: a map from each eigenvalue to a {-1,1}-basis for the corresponding eigenspace. (This field is nothing if the Laplacian is not spectrum integral.)

Examples

TODO: Add examples

Notes

laplacian_spectra_01neg outsources the actual computation logic to the _typed_laplacian_spectra_01neg helper, which dispatches to separate methods for different categories of Laplacians. The input matrix is first classified and cast to a concrete subtype of _TypedLaplacian, allowing said helper to apply specialized algorithms relying on the different {-1,0,1}-spectral properties of different types of graphs. See the _TypedLaplacian documentation for more details on the different classifications of Laplacians and the _typed_laplacian_spectra_01neg documentation for more details on the optimized algorithm used for each type.

SDiagonalizability.s_bandwidth — Method

s_bandwidth()

TODO: Write here. Just a placeholder for Documenter.jl as of right now.

References

[Fox09]

J. Fox. Lecture 19: The Petersen graph and Moore graphs. Lecture notes, MAT 307: Combinatorics, Massachusetts Institute of Technology (2009).

[JP25]

N. Johnston and S. Plosker. Laplacian {−1,0,1}- and {−1,1}-diagonalizable graphs. Linear Algebra and its Applications 704, 309–39 (2025).

[Joy15]

D. Joyce. Rotations and complex eigenvalues. Lecture notes, Math 130: Linear Algebra, Clark University (2015).