SDiagonalizability.jl – Private API

Documentation for SDiagonalizability's private API.

Note

The following documentation covers only the private API of the package. For public details, see the public API documentation.

Warning

The following bindings are internal; they may change or be removed in future versions.

SDiagonalizability._ORTHOGONALITY — Constant

_ORTHOGONALITY::_Orthogonality

A singleton instance of the semantically unique _Orthogonality type.

This constant is used to represent the property of pairwise orthogonality for a collection of vectors and should always be used in favor of instantiating a new object with _Orthogonality() every time.

[TODO: Justify further]

SDiagonalizability._QUASI_ORTHOGONALITY — Constant

_QUASI_ORTHOGONALITY::_QuasiOrthogonality

A singleton instance of the semantically unique _QuasiOrthogonality type.

This constant is used to represents the property of quasi-orthogonality for a collection of vectors and should always be used in favor of instantiating a new object with _QuasiOrthogonality() every time.

[TODO: Justify further]

SDiagonalizability._ArbitraryGraphLaplacian — Type

struct _ArbitraryGraphLaplacian

A wrapper for the Laplacian of any graph on n ≥ 3 nodes with no particular structure of interest.

Fields

matrix::Matrix{Int}: the n-by-n Laplacian in question. (Necessarily symmetric with zero row sums for some n ≥ 3.)

Supertype Hierarchy

_ArbitraryGraphLaplacian <: _TypedLaplacian <: Any

Notes

The only graph on zero nodes is the null graph (handled by the _NullGraphLaplacian type), the only graph on one node is an empty graph (handled by the _EmptyGraphLaplacian type), and the only graphs on two nodes are an empty graph and a complete graph (handled by the _CompleteGraphLaplacian type), so we are confident in the assumption that n ≥ 3 for any _ArbitraryGraphLaplacian instance.

See the documentation for laplacian_spectra_01neg for further details on the {-1,0,1}-spectral properties of this Laplacian type.

See also the documentation for parent type _TypedLaplacian if seeking justification for typing the matrix field as Matrix{Int} rather than AbstractMatrix{<:Integer}.

SDiagonalizability._CompleteGraphLaplacian — Type

struct _CompleteGraphLaplacian

A wrapper for the Laplacian of any uniformly weighted complete graph on n ≥ 2 nodes.

Fields

matrix::Matrix{Int}: the n-by-n Laplacian in question. (Necessarily symmetric with zero row sums and all off-diagonal entries equal to -weight for some n ≥ 2.)

weight::Int: the weight of each edge in the graph represented by matrix. (Necessarily a nonzero integer.)

Supertype Hierarchy

_CompleteGraphLaplacian <: _TypedLaplacian <: Any

Notes

The only graph on zero nodes is the null graph (handled by the _NullGraphLaplacian type) and the only graph on one node is an empty graph (handled by the _EmptyGraphLaplacian type), so we are confident in the assumption that n ≥ 2 for any _CompleteGraphLaplacian instance.

See the documentation for laplacian_spectra_01neg for further details on the {-1,0,1}-spectral properties of this Laplacian type.

See also the documentation for parent type _TypedLaplacian if seeking justification for typing the matrix field as Matrix{Int} rather than AbstractMatrix{<:Integer}.

SDiagonalizability._Eigenspace01Neg — Type

struct _Eigenspace01Neg

TODO: Write here

SDiagonalizability._EmptyGraphLaplacian — Type

struct _EmptyGraphLaplacian

A wrapper for the Laplacian matrix of any graph with no edges on n ≥ 1 nodes.

Fields

matrix::Matrix{Int}: the n-by-n Laplacian in question. (Necessarily equal to zeros(Int, n, n) for some n ≥ 1.)

Supertype Hierarchy

_EmptyGraphLaplacian <: _TypedLaplacian <: Any

Notes

We are confident in the assumption that n is at least 1 because the only graph on zero nodes is the null graph, which is handled by the _NullGraphLaplacian type.

See the documentation for laplacian_spectra_01neg for further details on the {-1,0,1}-spectral properties of this Laplacian type.

See also the documentation for parent type _TypedLaplacian if seeking justification for typing the matrix field as Matrix{Int} rather than AbstractMatrix{<:Integer}.

SDiagonalizability._KOrthogonality — Type

abstract type _KOrthogonality

An abstract type representing the property of k-orthogonality of a collection of vectors.

Recall that an (ordered) collection of vectors v₁, v₂, ..., vₙ is said to be k-orthogonal if we have the inner product ⟨vᵢ, vⱼ⟩ = 0 whenever |i - j| ≥ k (i.e., if every pair of vectors at least k indices apart is orthogonal). This is equivalent to the vectors' Gram matrix having bandwidth at most k, where we define the bandwidth of a matrix A to be the minimum integer k ≥ 1 such that aᵢⱼ = 0 whenever |i - j| ≥ k [JP25, p. 313]. (Note that many texts instead define matrix bandwidth using zero-based indexing—that is, with the condition |i - j| > k [Maf14, p. 186].)

This type is used as a template for concretely defined properties corresponding to specific values of k. For each family of values, we apply a separate algorithm to determine whether there exists a permutation of a given collection of vectors that induces k-orthogonality.

Interface

Concrete subtypes of _KOrthogonality must implement the following properties:

k::Int: the k-orthogonality parameter.

k need not necessarily be a field of the subtype (especially in the case of single-value types like _Orthogonality and _QuasiOrthogonality, in which case making k a field may erroneously imply that it is variable), but it must be accessible via Base.getproperty.

Subtypes

_Orthogonality: k-orthogonality for k = 1.

_QuasiOrthogonality: k-orthogonality for k = 2.

_WeakOrthogonality: k-orthogonality for k > 2.

Notes

Given the domain restriction of k to the positive integers, combined with the relatively small nature (≤ 25) of bandwidths examined by this package's principal S-bandwidth minimization algorithm, it may seem prudent to ask why we do not type the k field as UInt8 instead of Int. An equally compelling argument can be made to type it as Integer for genericity as well. [TODO: Justify]

On that note, the stipulation in the interface contract that k be an accessible property even for _Orthogonality and _QuasiOrthogonality [TODO: Elaborate]

Another design choice worth scrutinizing is the "_WeakOrthogonality" name. "Orthogonality" is a term as old as time and "quasi-orthogonality" too has risen to prominence in recent years (e.g., [JP25, p. 313]), but "weak orthogonality," strictly speaking, is not standard terminology. We have taken the liberty of coining it here for k > 2 to emphasize its weaker nature compared to orthogonality and quasi-orthogonality. That said, we make no attempts to formally introduce as a new definition in the broader literature beyond its ad hoc use (and but internally, at that) in this module.

Finally, we note that it may be slightly misleading to state that this type is used to inform our choice of algorithm in determining whether a collection of vectors is k-orthogonalizable (i.e., whether there exists a k-orthogonal permutation of said collection). Rather, we use it when determining whether a large, linearly dependent set of vectors contains an k-orthogonal independent spanning subset. This is still related to the long-standing matrix bandwidth minimization problem, but it differs in that we can take advantage of dynamic programming techniques when recursively searching for such a subset. (See the documentation for _find_k_orthogonal_basis for further details.)

SDiagonalizability._NullGraphLaplacian — Type

struct _NullGraphLaplacian

A wrapper for the Laplacian matrix of the (unique) graph with no nodes.

Fields

matrix::Matrix{Int}: the Laplacian in question. (Necessarily a 0×0 Matrix{Int}.)

Supertype Hierarchy

_NullGraphLaplacian <: _TypedLaplacian <: Any

Notes

See the documentation for laplacian_spectra_01neg for further details on the {-1,0,1}-spectral properties of this Laplacian type.

See also the documentation for parent type _TypedLaplacian if seeking justification for typing the matrix field as Matrix{Int} rather than AbstractMatrix{<:Integer}.

SDiagonalizability._Orthogonality — Type

struct _Orthogonality

Represents the property of pairwise orthogonality for a collection of vectors.

Recall that a collection of vectors v₁, v₂, ..., vₙ is said to be pairwise orthogonal if we have the inner product ⟨vᵢ, vⱼ⟩ = 0 whenever |i - j| ≥ 1 (i.e., if every pair of vectors is orthogonal). This is equivalent to the vectors' Gram matrix being diagonal.

Properties

k::Int: the k-orthogonality parameter. (Necessarily 1.)

Supertype Hierarchy

_Orthogonality <: _KOrthogonality <: Any

Notes

Since this type is semantically unique inasmuch as it simply encodes the information k = 1, we define the singleton instance _ORTHOGONALITY to always be used when needed in favor of instantiating an identical object every time.

SDiagonalizability._QOBasisSearchNodeData — Type

mutable struct _QOBasisSearchNodeData

TODO: Write here

SDiagonalizability._QuasiOrthogonality — Type

struct _QuasiOrthogonality

Represents the property of quasi-orthogonality for a collection of vectors.

Recall that an (ordered) collection of vectors v₁, v₂, ..., vₙ is said to be quasi-orthogonal if we have the inner product ⟨vᵢ, vⱼ⟩ = 0 whenever |i - j| ≥ 2 (i.e., if every pair of vectors at least 2 indices apart is orthogonal). This is equivalent to the vectors' Gram matrix being tridiagonal [JP25, p. 313].

Properties

k::Int: the k-orthogonality parameter. (Necessarily 2.)

Supertype Hierarchy

_QuasiOrthogonality <: _KOrthogonality <: Any

Notes

Since this type is semantically unique inasmuch as it simply encodes the information k = 2, we define the singleton instance _QUASI_ORTHOGONALITY to always be used when needed in favor of instantiating an identical object every time.

SDiagonalizability._TypedLaplacian — Type

abstract type _TypedLaplacian

An abstract type representing a classified Laplacian matrix of an undirected graph.

Interface

Concrete subtypes of _TypedLaplacian` must implement the following fields:

matrix::Matrix{Int}: the Laplacian matrix of the graph.

Subtypes

_NullGraphLaplacian`: a Laplacian wrapper for the (unique) graph with no nodes.

_EmptyGraphLaplacian: a Laplacian wrapper for any graph with no edges on n ≥ 1 nodes.

_CompleteGraphLaplacian: a Laplacian wrapper for any graph on n ≥ 2 nodes where every pair of nodes is connected by an edge and all edges possess the same weight.

_ArbitraryGraphLaplacian: a Laplacian wrapper for any graph on n ≥ 3 nodes with no particular strcuture of relevance in the context of investigating {-1,0,1}-spectra.

Notes

Given that this type and all its subtypes are privately encapsulated, we expect _TypedLaplacian instances to be created exclusively via the _cast_to_typed_laplacian factory function, which copies each input matrix to avoid shared mutability and converts it to a Matrix{Int} in the process. Therefore, it is safe to restrict the type of matrix field (in each concrete subtype) to Matrix{Int} in lieu of the more generic AbstractMatrix{<:Integer}. This decision may contribute marginal performance benefits via reduced type complexity and is highly unlikely to cause any issues (as this type is never to be exported to the public API). It also serves as an additional safeguard against directly passing a (wrongly typed) matrix to a subtype's constructor rather than calling _cast_to_typed_laplacian.

On a related note, it seems prudent to discuss the intended use case of _TypedLaplacian. The interface and its subtypes are meant to support _cast_to_typed_laplacian. This factory, in turn, determines which algorithm laplacian_spectra_01neg is to implement when computing data on the Laplacian matrix of a given graph, as different types of graphs exhibit different spectral properties which we may exploit to improve performance.

SDiagonalizability._WeakOrthogonality — Type

struct _WeakOrthogonality

Represents the property of "weak orthogonality" for a collection of vectors.

In particular, "weak orthogonality" is an ad hoc term used to refer to the property of k-orthogonality for k > 2. Recall that an (ordered) collection of vectors v₁, v₂, ..., vₙ is said to be k-orthogonal if we have the inner product ⟨vᵢ, vⱼ⟩ = 0 whenever |i - j| ≥ k (i.e., if every pair of vectors at least k indices apart is orthogonal). This is equivalent to the vectors' Gram matrix having bandwidth at most k.

Fields

k::Int: the k-orthogonality parameter. (Necessarily a positive integer greater than 2.)

Supertype Hierarchy

_WeakOrthogonality <: _KOrthogonality <: Any

Constructors

_WeakOrthogonality(k::Int): constructs a _WeakOrthogonality object with the given k-orthogonality parameter. Throws a DomainError if k <= 2.

Notes

The term "weak orthogonality" is not standard terminology in the literature, but it is used here to emphasize the weaker nature of this property compared to orthogonality and quasi-orthogonality. It is an ad hoc term coined for this module and is not intended to be formally introduced in the broader literature. See also the documentation for parent type _KOrthogonality for further discussion of this topic.

SDiagonalizability._assert_matrix_is_undirected_laplacian — Method

_assert_matrix_is_undirected_laplacian(L)

Validate that L is the Laplacian matrix of an undirected, possibly weighted graph.

Arguments

L::AbstractMatrix{<:Integer}: the purported Laplacian matrix.

Returns

nothing: if the check is passed, no output is produced.

Throws

DomainError: if L is not symmetric or has nonzero row sums.

Examples

The Laplacian matrix of the (undirected) star graph on 5 vertices passes the check:

julia> L = [ 4  -1  -1  -1  -1;
            -1   1   0   0   0;
            -1   0   1   0   0;
            -1   0   0   1   0;
            -1   0   0   0   1]
5×5 Matrix{Int64}:
  4  -1  -1  -1  -1
 -1   1   0   0   0
 -1   0   1   0   0
 -1   0   0   1   0
 -1   0   0   0   1

julia> isnothing(SDiagonalizability._assert_matrix_is_undirected_laplacian(L))
true

The adjacency matrix of the (undirected) cycle graph on 4 vertices is symmetric but has nonzero row sums, so it fails the check:

julia> A = [0  1  0  1;
            1  0  1  0;
            0  1  0  1;
            1  0  1  0]
4×4 Matrix{Int64}:
 0  1  0  1
 1  0  1  0
 0  1  0  1
 1  0  1  0

julia> SDiagonalizability._assert_matrix_is_undirected_laplacian(A)
ERROR: DomainError with [0 1 0 1; 1 0 1 0; 0 1 0 1; 1 0 1 0]:
Matrix has nonzero row sums; cannot be an (undirected) Laplacian
[...]

Both the in-degree and out-degree Laplacian matrices of this random tournament digraph have zero row sums but are not symmetric, so they fail the check. (These are the two standard ways of extending the concept of the Laplacian to directed graphs [VL20, p. 196].)

julia> G = random_tournament_digraph(3; seed=87)
{3, 3} directed simple Int64 graph

julia> L_in = laplacian_matrix(G; dir=:in)
3×3 SparseArrays.SparseMatrixCSC{Int64, Int64} with 5 stored entries:
  ⋅  ⋅   ⋅
 -1  2  -1
 -1  ⋅   1

julia> SDiagonalizability._assert_matrix_is_undirected_laplacian(L_in)
ERROR: DomainError with sparse([2, 3, 2, 2, 3], [1, 1, 2, 3, 3], [-1, -1, 2, -1, 1], 3, 3):
Matrix is not symmetric; cannot be an (undirected) Laplacian
[...]

julia> L_out = laplacian_matrix(G; dir=:out)
3×3 SparseArrays.SparseMatrixCSC{Int64, Int64} with 5 stored entries:
 2  -1  -1
 ⋅   ⋅   ⋅
 ⋅  -1   1

julia> SDiagonalizability._assert_matrix_is_undirected_laplacian(L_out)
ERROR: DomainError with sparse([1, 1, 3, 1, 3], [1, 2, 2, 3, 3], [2, -1, -1, -1, 1], 3, 3):
Matrix is not symmetric; cannot be an (undirected) Laplacian
[...]

Notes

If edges are to be bidirectional, then L must be symmetric. L must also have zero row sums, since the (i, i)-th entry is the weighted degree of node i (the sum of all incident edges' weights) and the (i, j)-th entry for i ≠ j is the negation of the weight of edge (i, j) (or simply 0, if no such edge exists).

Given the highly optimized, lazy, zero-allocation implementation of LinearAlgebra.issymmetric, the symmetry check is performed first. (Both steps are O(n²) in the worst case, but testing for symmetry is far more performant in practice.) This also allows us to (also lazily) check for nonzero column sums rather than nonzero row sums (since these are equivalent for symmetric matrices) in the second step, taking advantage of Julia's column-major storage model.

At first blush, it may seem as though the choice of DomainError over something like ArgumentError (or even simply the return of a boolean) constitutes poor design. However, this is informed by the simple ad hoc use of this function to validate inputs for other functions requiring Laplacian matrices. Certainly, this function is never meant to be publicly exposed on its own.

SDiagonalizability._cast_to_typed_laplacian — Method

_cast_to_typed_laplacian(L)

Classify the Laplacian matrix L and wrap it in a _TypedLaplacian object.

It is first verified that L is indeed a Laplacian matrix by _assert_matrix_is_undirected_laplacian, which throws a DomainError otherwise. It is then classified based on any properties which may be exploited in computing data on its {-1,0,1}-spectrum.

Arguments

L::AbstractMatrix{<:Integer}: the Laplacian matrix to classify.

Returns

TL::_TypedLaplacian: the Laplacian wrapped in a concrete _TypedLaplacian subtype associated with the category of the graph represented by L. In the case of complete graphs, TL also contains data on the (necessarily uniform) weight of each edge.

Examples

Correctly recognizes the Laplacian matrix of the null graph:

julia> G = Graph(0)
{0, 0} undirected simple Int64 graph

julia> L = laplacian_matrix(G)
0×0 SparseArrays.SparseMatrixCSC{Int64, Int64} with 0 stored entries

julia> TL = SDiagonalizability._cast_to_typed_laplacian(L)
SDiagonalizability._NullGraphLaplacian(Matrix{Int64}(undef, 0, 0))

julia> isa(TL, SDiagonalizability._NullGraphLaplacian)
true

Correctly recognizes the Laplacian matrix of the empty graph on 3 nodes:

julia> G = Graph(3)
{3, 0} undirected simple Int64 graph

julia> L = laplacian_matrix(G)
3×3 SparseArrays.SparseMatrixCSC{Int64, Int64} with 0 stored entries:
 ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅

julia> TL = SDiagonalizability._cast_to_typed_laplacian(L)
SDiagonalizability._EmptyGraphLaplacian([0 0 0; 0 0 0; 0 0 0])

julia> isa(TL, SDiagonalizability._EmptyGraphLaplacian)
true

Correctly recognizes the Laplacian matrix of the complete graph on 4 nodes:

julia> G = complete_graph(4)
{4, 6} undirected simple Int64 graph

julia> L = laplacian_matrix(G)
4×4 SparseArrays.SparseMatrixCSC{Int64, Int64} with 16 stored entries:
  3  -1  -1  -1
 -1   3  -1  -1
 -1  -1   3  -1
 -1  -1  -1   3

julia> TL = SDiagonalizability._cast_to_typed_laplacian(L)
SDiagonalizability._CompleteGraphLaplacian([3 -1 -1 -1; -1 3 -1 -1; -1 -1 3 -1; -1 -1 -1 3], -1)

julia> isa(TL, SDiagonalizability._CompleteGraphLaplacian)
true

Correctly recognizes the Laplacian matrix of this random generic graph:

julia> G = erdos_renyi(5, 8; seed=87)
{5, 8} undirected simple Int64 graph

julia> L = laplacian_matrix(G)
5×5 SparseArrays.SparseMatrixCSC{Int64, Int64} with 21 stored entries:
  3   ⋅  -1  -1  -1
  ⋅   2  -1  -1   ⋅
 -1  -1   4  -1  -1
 -1  -1  -1   4  -1
 -1   ⋅  -1  -1   3

julia> TL = SDiagonalizability._cast_to_typed_laplacian(L)
SDiagonalizability._ArbitraryGraphLaplacian([3 0 … -1 -1; 0 2 … -1 0; … ; -1 -1 … 4 -1; -1 0 … -1 3])

julia> isa(TL, SDiagonalizability._ArbitraryGraphLaplacian)
true

Notes

Right now, the possible types of Laplacian matrices (or, to be more precise, the graphs they represent) are:

_NullGraphLaplacian: the (unique) graph with no nodes.

_EmptyGraphLaplacian: any graph with no edges on n ≥ 1 nodes.

_CompleteGraphLaplacian: any graph on n ≥ 2 nodes where every pair of nodes is connected by an edge and all edges possess the same weight.

_ArbitraryGraphLaplacian: any graph on n ≥ 3 nodes with no particular strcuture of relevance in the context of investigating {-1,0,1}-spectra.

SDiagonalizability._classify_orthogonality_property — Method

_classify_orthogonality_property(k)

Classifies the k-orthogonality property based on the given k parameter.

When searching for a k-orthogonal S-basis of a given Laplacian eigenspace, the family of values to which our k parameter belongs informs our choice of algorithm.

Arguments

k::Int: the k-orthogonality parameter to classify. Must be a positive integer.

Returns

::_KOrthogonality: An instance of a concrete _KOrthogonality subtype corresponding to k.

Throws

DomainError: if k is not a positive integer.

Examples

julia> SDiagonalizability._classify_orthogonality_property(1)
SDiagonalizability._Orthogonality()

julia> SDiagonalizability._classify_orthogonality_property(2)
SDiagonalizability._QuasiOrthogonality()

julia> SDiagonalizability._classify_orthogonality_property(3)
SDiagonalizability._WeakOrthogonality(3)

julia> SDiagonalizability._classify_orthogonality_property(13)
SDiagonalizability._WeakOrthogonality(13)

Notes

See the documentation for _find_k_orthogonal_basis for further details on how this function is used in the context of finding a k-orthogonal basis.

SDiagonalizability._extract_independent_cols — Method

_extract_independent_cols(A)

Return a (not necessarily unique) independent spanning subset of the columns of A.

Computing a QR decomposition of A, the scaling coefficients from the orthogonalization process are used to determine the rank (rather than recompute it with an SVD), while the pivots are used to extract a spanning set of independent columns.

Arguments

A::AbstractMatrix{T<:Integer}: the matrix whose independent columns to extract.

Returns

::AbstractMatrix{T}: a spanning set of independent columns of A.

Examples

Observe how columns with greater Euclidean norms are given priority in the pivot ordering:

julia> A = [3  0  0  0  2  1   5   0
            0  3  0  0  2  1  -5   0
            0  0  3  0  2  1   5   4
            0  0  0  3  2  1   0  -4
            0  0  0  0  0  0   0   0]
5×8 Matrix{Int64}:
 3  0  0  0  2  1   5   0
 0  3  0  0  2  1  -5   0
 0  0  3  0  2  1   5   4
 0  0  0  3  2  1   0  -4
 0  0  0  0  0  0   0   0

julia> SDiagonalizability._extract_independent_cols(A)
5×4 Matrix{Int64}:
  5   0  2  3
 -5   0  2  0
  5   4  2  0
  0  -4  2  0
  0   0  0  0

Notes

The somewhat rare (although certainly far from unorthodox) choice of QR decomposition for our purposes merits discussion. [TODO: Write here]

TODO: Also discuss why we can't just transpose first (gets us independent rows)

TODO: Discuss unusual use of _rank_rtol

SDiagonalizability._find_basis_indices — Method

_find_basis_indices(
    curr_indices,
    column_space,
    column_rank,
    G::Graph,
    H::Graph,
    prop::_WeakOrthogonality,
)

TODO: Write here

SDiagonalizability._find_basis_indices — Method

_find_basis_indices(curr_indices, column_space, column_rank, prop::_Orthogonality)

TODO: Write here

SDiagonalizability._find_basis_indices — Method

_find_basis_indices(
    curr_indices,
    column_space,
    column_rank,
    nodes,
    union_find,
    prop::_QuasiOrthogonality,
)

TODO: Write here

SDiagonalizability._find_basis_with_property — Method

_find_basis_with_property(column_space, column_rank, prop::_KOrthogonality)

TODO: Write here

SDiagonalizability._find_basis_with_property — Method

_find_basis_with_property(column_space, column_rank, prop:::_Orthogonality)

TODO: Write here

SDiagonalizability._find_basis_with_property — Method

_find_basis_with_property(column_space, column_rank, prop::_QuasiOrthogonality)

TODO: Write here

SDiagonalizability._find_basis_with_property — Method

_find_basis_with_property(column_space, column_rank, prop::_WeakOrthogonality)

TODO: Write here

SDiagonalizability._find_indices_1neg — Method

_find_indices_1neg(eigvecs_01neg)

Find the indices of {-1,1}-eigenvectors given a map of eigenspaces.

More precisely, given a map from each eigenvalue to all {-1,0,1}-vectors in the associated eigenspace, this function returns a map from each eigenvalue to the indices of the {-1,1}-vectors in each value from the input map.

Arguments

eigvecs_01neg::AbstractDict{Int,<:AbstractMatrix{Int}}: a map from eigenvalues to eigenspaces, TODO: Write here

Returns

::AbstractDict{Int,Vector{Int}}: TODO: Write here

Examples

TODO: Write here

Notes

TODO: Justify decision to use Vector{Int} rather than BitVector for the indices

SDiagonalizability._find_k_orthogonal_basis — Method

_find_k_orthogonal_basis(column_space, column_rank, k)

TODO: Write here

SDiagonalizability._pot_kernel_eigvecs_01neg — Method

_pot_kernel_eigvecs_01neg(n)

Lazily generate all potential kernel {-1,0,1}-eigenvectors of an order n Laplacian.

Each vector is normalized so that its first nonzero entry is 1, enforcing pairwise linear independence between all generated vectors.

Arguments

n::Integer: the order of the Laplacian matrix of some undirected graph for which to find potential kernel eigenvectors.

Returns

eigvec_generator::Iterators.Flatten{<:Base.Generator}: a lazily evaluated iterator over all {-1,0,1}-vectors in ℝⁿ, unique up to span.

Throws

DomainError: if n is negative.

Examples

Generate all potential kernel eigenvectors for an order 3 Laplacian matrix:

julia> hcat(SDiagonalizability._pot_kernel_eigvecs_01neg(3)...)
3×13 Matrix{Int64}:
  1   1   1   1  1  1   1  1  1   0  0  0  0
 -1  -1  -1   0  0  0   1  1  1   1  1  1  0
 -1   0   1  -1  0  1  -1  0  1  -1  0  1  1

Notes

The number of potential kernel eigenvectors (unique up to span) for an order n Laplacian matrix is given by (3ⁿ - 1) / 2. See also the relevant OEIS sequence [Slo25].

Regrettably, the implementation here is rather clunky and unidiomatic, but it is worth noting that eigenvector generation (and subsequent processing by laplacian_spectra_01neg) is one of two major bottlenecks in the overall S-bandwidth minimization algorithm. Given how much potential there is for optimization in this piece of code, we thus prioritize performance over readability in this particular case, making every effort to include inline comments wherever clarification may be needed.

SDiagonalizability._pot_nonkernel_eigvecs_01neg — Method

_pot_nonkernel_eigvecs_01neg(n)

Lazily generate all potential non-kernel {-1,0,1}-eigenvectors of an order n Laplacian.

Each vector is normalized so that its first nonzero entry is 1, enforcing pairwise linear independence between all generated vectors. Since all Laplacian matrices have pairwise orthogonal eigenspaces and the all-ones vector is always in the kernel, every non-kernel {-1,0,1}-eigenvector must also have an equal number of -1's and 1's.

Arguments

n::Integer: the order of the Laplacian matrix of some undirected graph for which to find potential non-kernel eigenvectors.

Returns

eigvec_generator::Iterators.Flatten{<:Base.Generator}: a lazily evaluated iterator over all {-1,0,1}-vectors in ℝⁿ orthogonal to the all-ones kernel vector, unique up to span.

Throws

DomainError: if n is negative.

Examples

Generate all potential non-kernel eigenvectors of an order 4 Laplacian matrix:

julia> hcat(SDiagonalizability._pot_nonkernel_eigvecs_01neg(4)...)
4×9 Matrix{Int64}:
  1   1   1   1   1   1   0   0   0
 -1   0   0  -1  -1   1   1   1   0
  0  -1   0  -1   1  -1  -1   0   1
  0   0  -1   1  -1  -1   0  -1  -1

Notes

The number of potential non-kernel eigenvectors (unique up to span) for an order n Laplacian matrix is, by non-trivial combinatorial arguments, equal to the number of humps in all Motzkin paths of length n. See also the relevant OEIS sequence [Deu25].

Regrettably, the implementation here is rather clunky and unidiomatic, but it is worth noting that eigenvector generation (and subsequent processing by laplacian_spectra_01neg) is one of two major bottlenecks in the overall S-bandwidth minimization algorithm. Given how much potential there is for optimization in this piece of code, we thus prioritize performance over readability in this particular case, making every effort to include inline comments wherever clarification may be needed.

SDiagonalizability._rank_rtol — Method

function _rank_rtol(A::AbstractMatrix{<:Real})

Return a reasonable relative tolerance for computing matrix rank via SVD or QRD.

If using SVD to compute rank with singular values Ω, then the intended use case is to numerically estimate the rank r of A by r = count(Ω .> maximum(Ω) * _rank_rtol(A)). Similarly, if using QRD to compute rank with scaling coefficients Δ (the diagonal entries of the R matrix in A = QR), then the user should numerically estimate the rank by r = count(Δ .> maximum(Δ) * _rank_rtol(A)).

Since the machine epsilon is only defined for floating-point types, we call eps() (defaulting to Float64) instead of eps(eltype(A)) when computing the tolerance. If the elements of A are floats, on the other hand, then this function dispatches to a separate method that uses the element type's machine epsilon.

Arguments

A::AbstractMatrix{<:Real}: the matrix for which to compute a tolerance.

Returns

tol::Float64: a reasonable relative tolerance for computing matrix rank via SVD or QRD. This scales proportionally to the maximum dimension of A.

Notes

This is very much an ad hoc function applied specifically for numerical stability within this package and is never meant to be used as part of the public API. The sqrt(maximum(size(A)) * eps()) formula is inspired by NumPy's default relative tolerance of maximum(size(A)) * eps() in the numpy.linalg.matrix_rank function [Num25], with a square root taken for additional robustness. (In particular, the short, fat matrices holding all {-1,0,1}-eigenvectors of a Laplacian matrix are often ill-conditioned, hence the need for this higher tolerance.)

In the case of using SVD to compute matrix rank, the decision to default to the Float64 machine epsilon is motivated by how LinearAlgebra.rank handles non-floating-point matrices, as LAPACK automatically converts to Float64 under the hood. In the case of QRD, our ad hoc use of this function in _extract_independent_cols takes in floating-point matrices anyway, so this particular method is no longer relevant.

SDiagonalizability._rank_rtol — Method

function _rank_rtol(A::AbstractMatrix{<:AbstractFloat})

Return a reasonable relative tolerance for computing matrix rank via SVD or QRD.

If using SVD to compute rank with singular values Ω, then the intended use case is to numerically estimate the rank r of A by r = count(Ω .> maximum(Ω) * _rank_rtol(A)). Similarly, if using QRD to compute rank with scaling coefficients Δ (the diagonal entries of the R matrix in A = QR), then the user should numerically estimate the rank by r = count(Δ .> maximum(Δ) * _rank_rtol(A)).

Arguments

A::AbstractMatrix{T<:AbstractFloat}: the matrix for which to compute a tolerance.

Returns

tol::Float64: a reasonable relative tolerance for computing matrix rank via SVD or QRD. This scales proportionally to the maximum dimension of A and the machine epsilon of the element type T.

Notes

This is very much an ad hoc function applied specifically for numerical stability within this package and is never meant to be used as part of the public API. The sqrt(maximum(size(A)) * eps()) formula is inspired by NumPy's default relative tolerance of maximum(size(A)) * eps() in the numpy.linalg.matrix_rank function [Num25], with a square root taken for additional robustness. (In particular, the short, fat matrices holding all {-1,0,1}-eigenvectors of a Laplacian matrix are often ill-conditioned, hence the need for this higher tolerance.)

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_ArbitraryGraphLaplacian)

TODO: Write here

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_CompleteGraphLaplacian)

TODO: Write here

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_EmptyGraphLaplacian)

TODO: Write here

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_NullGraphLaplacian)

TODO: Write here

SDiagonalizability._typed_laplacian_spectra_01neg — Method

_typed_laplacian_spectra_01neg(TL::_TypedLaplacian)

TODO: Write here

References

[Deu25]

E. Deutsch. Number of humps in all Motzkin paths of length n. Entry A097861 in The On-Line Encyclopedia of Integer Sequences (2025). Accessed: 2025-05-22.

[Maf14]

L. O. Mafteiu-Scai. The Bandwidths of a Matrix. A Survey of Algorithms. Annals of West University of Timisoara - Mathematics and Computer Science 52, 183–223 (2014).

[Slo25]

N. J. Sloane, a(n) = (3^n - 1)/2. Entry A003462 in The On-Line Encyclopedia of Integer Sequences (2025). Accessed: 2025-05-22.

[VL20]

J. J. Veerman and R. Lyons. A Primer on Laplacian Dynamics in Directed Graphs. Nonlinear Phenomena in Complex Systems 23, 196–206 (2020).

[Num25]

NumPy Developers, numpy.linalg.matrix_rank. NumPy reference documentation (stable) (2025). Accessed: 2025-05-22.