SDiagonalizability.jl – Private API

Documentation for SDiagonalizability's private API.

Note

The following documentation covers only the private API of the package. For public details, see the public API documentation.

SDiagonalizability.AbstractSBandResult — Type

AbstractSBandResult

Abstract base type for all S-bandwidth problem results.

Interface

Concrete subtypes of AbstractSBandResult must implement parametric types

A<:Union{AbstractGraph,AbstractMatrix{<:Integer}};

B<:Tuple; and

C<:Union{Nothing,Eigen},

alongside the following fields:

network::A: the network whose S-bandwidth is investigated;

S::B: the set S from whose entries we are allowed to construct eigenvectors;

diagonalization::C: an S-diagonalization of the matrix representation of the network, if it satisfies the specified S-bandwidth constraints; otherwise, nothing.

SDiagonalizability.EfficiencyWarning — Type

EfficiencyWarning(msg)

[TODO: Write here]

SDiagonalizability.NotImplementedError — Type

NotImplementedError{Nothing}(f, subtype, abstracttype)
NotImplementedError{Symbol}(f, arg, subtype, abstracttype)

An exception indicating that a function lacks dispatch to handle a specific argument type.

Semantically, this differs from MethodError in that it connotes a developer-side failure to implement a method rather than erroneous user input. Throughout this package, it is often used to warn when an existing function with multiple dispatch on some abstract type is called on a newly created subtype for which no method has been defined.

Fields

f::Function: the function called.

arg::Symbol: the name of the argument with the unsupported type, if the function has multiple arguments. If the function has only one argument, this field should be set to nothing.

subtype::Type: the type of the argument. May be the actual concrete type or some intermediate supertype. (For instance, if the relevant input has concrete type A with hierarchy A <: B <: C and the abstracttype field is C, then both A and B are perfectly valid choices for subtype.)

abstracttype::Type: the abstract type under which the argument is meant to fall.

Constructors

NotImplementedError(::Function, ::Type, ::Type): constructs a new NotImplementedError instance for a single-argument function. Throws an error if the second type is not abstract or the first type is not a subtype of the second.

NotImplementedError(::Function, ::Symbol, ::Type, ::Type): constructs a new NotImplementedError instance for a multi-argument function. Throws an error if the second type is not abstract or the first type is not a subtype of the second.

SDiagonalizability.SBandMinimizationResult — Type

SBandMinimizationResult{A,B,C,D} <: AbstractSBandResult

[TODO: Write here]

SDiagonalizability.SBandRecognitionResult — Type

SBandRecognitionResult{A,B,C} <: AbstractSBandResult

[TODO: Write here]

SDiagonalizability.SpectrumIntegralResult — Type

struct SpectrumIntegralResult

Data on whether a matrix is spectrum integral (i.e., whether its eigenvalues are integers).

This struct also contains a map from each eigenvalue to its multiplicity, provided that the eigenvalues are indeed all integers. (Otherwise, the associated field is simply nothing.)

Fields

matrix::Matrix{Int}: the matrix whose eigenvalues and their integrality are of interest.

spectrum_integral::Bool: whether the eigenvalues of matrix are all integers.

multiplicities::Union{Nothing,OrderedDict{Int,Int}}: a map from each eigenvalue to its multiplicity, sorted first by ascending multiplicity then by ascending eigenvalue. (This field is nothing if and only if spectrum_integral is false, since we cannot map non-integer eigenvalues to data.)

Notes

If an undirected graph with integer edge weights is {-1,0,1}-diagonalizable (or, more restrictively, {-1,1}-diagonalizable), then its Laplacian matrix has integer eigenvalues [JP25, p. 312]. Hence, validating Laplacian integrality serves as a useful screening step in this package's principal S-bandwidth minimization algorithm.

SDiagonalizability._SSpectra — Type

_SSpectra{A,B,C,D}

[TODO: Write here]

LinearAlgebra.rank — Function

rank(A::QRPivoted{<:Any, T}; atol::Real=0, rtol::Real=min(n,m)*ϵ) where {T}

Compute the numerical rank of the QR factorization A by counting how many diagonal entries of A.factors are greater than max(atol, rtol*Δ₁) where Δ₁ is the largest calculated such entry. This is similar to the rank(::AbstractMatrix) method insofar as it counts the number of (numerically) nonzero coefficients from a matrix factorization, although the default method uses an SVD instead of a QR factorization. Like rank(::SVD), this method also re-uses an existing matrix factorization.

Computing rank via QR factorization should almost always produce the same results as via SVD, although this method may be more prone to overestimating the rank in pathological cases where the matrix is ill-conditioned. It is also worth noting that it is generally faster to compute a QR factorization than an SVD, so this method may be preferred when performance is a concern.

atol and rtol are the absolute and relative tolerances, respectively. The default relative tolerance is n*ϵ, where n is the size of the smallest dimension of A and ϵ is the eps of the element type of A.

SDiagonalizability._assert_matrix_is_undirected_laplacian — Method

_assert_matrix_is_undirected_laplacian(L)

Validate that L is the Laplacian matrix of an undirected, possibly weighted graph.

Arguments

L::AbstractMatrix{<:Integer}: the purported Laplacian matrix.

Returns

nothing: if the check is passed, no output is produced.

Throws

DomainError: if L is not symmetric or has nonzero row sums.

Examples

The Laplacian matrix of the (undirected) star graph on 5 vertices passes the check:

julia> L = [ 4  -1  -1  -1  -1;
            -1   1   0   0   0;
            -1   0   1   0   0;
            -1   0   0   1   0;
            -1   0   0   0   1]
5×5 Matrix{Int64}:
  4  -1  -1  -1  -1
 -1   1   0   0   0
 -1   0   1   0   0
 -1   0   0   1   0
 -1   0   0   0   1

julia> isnothing(SDiagonalizability._assert_matrix_is_undirected_laplacian(L))
true

The adjacency matrix of the (undirected) cycle graph on 4 vertices is symmetric but has nonzero row sums, so it fails the check:

julia> A = [0  1  0  1;
            1  0  1  0;
            0  1  0  1;
            1  0  1  0]
4×4 Matrix{Int64}:
 0  1  0  1
 1  0  1  0
 0  1  0  1
 1  0  1  0

julia> SDiagonalizability._assert_matrix_is_undirected_laplacian(A)
ERROR: DomainError with [0 1 0 1; 1 0 1 0; 0 1 0 1; 1 0 1 0]:
Matrix has nonzero row sums; cannot be an (undirected) Laplacian
[...]

Both the in-degree and out-degree Laplacian matrices of this random tournament digraph have zero row sums but are not symmetric, so they fail the check. (These are the two standard ways of extending the concept of the Laplacian to directed graphs [VL20, p. 196].)

julia> using Graphs

julia> G = random_tournament_digraph(3; seed=87)
{3, 3} directed simple Int64 graph

julia> L_in = laplacian_matrix(G; dir=:in)
3×3 SparseArrays.SparseMatrixCSC{Int64, Int64} with 5 stored entries:
  ⋅  ⋅   ⋅
 -1  2  -1
 -1  ⋅   1

julia> SDiagonalizability._assert_matrix_is_undirected_laplacian(L_in)
ERROR: DomainError with sparse([2, 3, 2, 2, 3], [1, 1, 2, 3, 3], [-1, -1, 2, -1, 1], 3, 3):
Matrix is not symmetric; cannot be an (undirected) Laplacian
[...]

julia> L_out = laplacian_matrix(G; dir=:out)
3×3 SparseArrays.SparseMatrixCSC{Int64, Int64} with 5 stored entries:
 2  -1  -1
 ⋅   ⋅   ⋅
 ⋅  -1   1

julia> SDiagonalizability._assert_matrix_is_undirected_laplacian(L_out)
ERROR: DomainError with sparse([1, 1, 3, 1, 3], [1, 2, 2, 3, 3], [2, -1, -1, -1, 1], 3, 3):
Matrix is not symmetric; cannot be an (undirected) Laplacian
[...]

Notes

If edges are to be bidirectional, then L must be symmetric. L must also have zero row sums, since the (i, i)-th entry is the weighted degree of node i (the sum of all incident edges' weights) and the (i, j)-th entry for i ≠ j is the negation of the weight of edge (i, j) (or simply 0, if no such edge exists).

Given the highly optimized, lazy, zero-allocation implementation of LinearAlgebra.issymmetric, the symmetry check is performed first. (Both steps are O(n²) in the worst case, but testing for symmetry is far more performant in practice.) This also allows us to (also lazily) check for nonzero column sums rather than nonzero row sums (since these are equivalent for symmetric matrices) in the second step, taking advantage of Julia's column-major storage model.

At first blush, it may seem as though the choice of DomainError over something like ArgumentError (or even simply the return of a boolean) constitutes poor design. However, this is informed by the simple ad hoc use of this function to validate inputs for other functions requiring Laplacian matrices. Certainly, this function is never meant to be publicly exposed on its own.

SDiagonalizability._rank_rtol — Method

function _rank_rtol(A::AbstractMatrix{<:Real})

Return a reasonable relative tolerance for computing matrix rank via SVD or QRD.

The output is intended to be passed as a keyword argument to LinearAlgebra.rank. LinearAlgebra.eigtype is used to determine the appropriate machine epsilon for the element type of A when eltype(A) is not an AbstractFloat.

Arguments

A::AbstractMatrix{<:Real}: the matrix for which to compute a tolerance.

Returns

tol::Float64: a reasonable relative tolerance for computing matrix rank via SVD or QRD. This scales proportionally to the maximum dimension of A.

Notes

LinearAlgebra.rank's default rtol of min(m,n) * ϵ for computing the rank of an m × n matrix may result in overestimating rank when |m - n| ≫ 0, since condition number (which determines how numerically stable SVD and QRD are) grows with both dimensions [CD05, p. 603]. Given that we often deal with short-and-fat matrices in this package (particularly when processing all {-1,0,1}-eigenvectors of a Laplacian matrix), we turn instead to the same relative tolerance used by NumPy's and MATLAB's rank functions—max(m,n) * ϵ [Num25, MAT25]. (Indeed, this is a widely adopted standard across the field of numerical analysis [PTVF07, p. 795].)

References

[CD05]

Z. Chen and J. Dongarra. Condition Numbers of Gaussian Random Matrices. SIAM Journal on Matrix Analysis and Applications 27, 603–20 (2005).

[JP25]

N. Johnston and S. Plosker. Laplacian {−1,0,1}- and {−1,1}-diagonalizable graphs. Linear Algebra and its Applications 704, 309–39 (2025).

[PTVF07]

W. H. Press, S. A. Teukolsky, W. T. Vetterling and B. P. Flannery. Numerical Recipes: The Art of Scientific Computing. 3rd Edition (Cambridge University Press, Cambridge, UK, 2007). ISBN: 978-0-521-88068-8.

[VL20]

J. J. Veerman and R. Lyons. A Primer on Laplacian Dynamics in Directed Graphs. Nonlinear Phenomena in Complex Systems 23, 196–206 (2020).

[MAT25]

MATLAB Developers, rank. MATLAB reference documentation – R2025a (2025). Accessed: 2025-05-29.

[Num25]

NumPy Developers, numpy.linalg.matrix_rank. NumPy reference documentation – v2.2 (2025). Accessed: 2025-05-22.