SDiagonalizability.jl – Public API

Documentation for SDiagonalizability's public API.

Note

The following documentation covers only the public API of the package. For internal details, see the private API documentation.

SDiagonalizability.SDiagonalizability — Module

SDiagonalizability

A dynamic algorithm to test a graph for S-diagonalizability and minimize its S-bandwidth.

SDiagonalizability.jl packages the first non-naïve deterministic algorithm to minimize the S-bandwidth of a quantum network (or, to be more precise, its graph representation), written in Julia. Given some finite S ∈ Zⁿ, the S-bandwidth of an (undirected) graph G with Laplacian L(G) ∈ Rⁿˣⁿ is the minimum integer k ≥ 1 such that L(G) = PDP⁻¹ for some diagonal D ∈ Rⁿˣⁿ and P ∈ Sⁿˣⁿ with [PᵀP]ᵢⱼ = 0 whenever |i - j| ≥ k. For specific choices of S (namely S = {-1,1} and S = {-1,0,1}), a quantum network's S-bandwidth has been shown to be an indicator of high state transfer fidelity due to automorphic properties of its graph representation.

Full documentation is available for the latest development version of this package.

SDiagonalizability.LaplacianSpectrum01Neg — Type

struct LaplacianSpectrum01Neg

TODO: Write here

SDiagonalizability.SpectrumIntegralResult — Type

struct SpectrumIntegralResult

TODO: Write here

SDiagonalizability.check_spectrum_integrality — Method

check_spectrum_integrality(A)

Check whether the eigenvalues of A are integers (up to floating-point error).

If the eigenvalues are indeed all integers, then an eigenvalue-multiplicity map is constructed as well.

Arguments

A::AbstractMatrix{<:Integer}: the matrix whose eigenvalues to check for integrality.

Returns

::SpectrumIntegralResult: a struct containing the following fields:

matrix::Matrix{Int}: a (casted) copy of A, avoiding shared mutability.

spectrum_integral::Bool: whether the eigenvalues of A are integers.

multiplicities::Union{Nothing,OrderedDict{Int,Int}}: a map from eigenvalues to multiplicities, sorted first by ascending multiplicity then by ascending eigenvalue. (This field is nothing if the eigenvalues are not all integers.)

Examples

Confirm that the rotation matrix by π/2 radians counterclockwise is not spectrum integral (rather, it has eigenvalues ±i [Joy15, p. 1]):

julia> R = Int8.([0 -1; 1 0])
2×2 Matrix{Int8}:
 0  -1
 1   0

julia> res = check_spectrum_integrality(R);

julia> res.matrix
2×2 Matrix{Int64}:
 0  -1
 1   0

julia> res.spectrum_integral
false

julia> isnothing(res.multiplicities)
true

Confirm that the adjacency matrix of the Petersen graph is spectrum integral, with correct eigenvalues and multiplicities of {3: 1, -2: 4, 1: 5} [Fox09, p. 2]:

julia> G = smallgraph(:petersen)
{10, 15} undirected simple Int64 graph

julia> A = adjacency_matrix(G)
10×10 SparseArrays.SparseMatrixCSC{Int64, Int64} with 30 stored entries:
 ⋅  1  ⋅  ⋅  1  1  ⋅  ⋅  ⋅  ⋅
 1  ⋅  1  ⋅  ⋅  ⋅  1  ⋅  ⋅  ⋅
 ⋅  1  ⋅  1  ⋅  ⋅  ⋅  1  ⋅  ⋅
 ⋅  ⋅  1  ⋅  1  ⋅  ⋅  ⋅  1  ⋅
 1  ⋅  ⋅  1  ⋅  ⋅  ⋅  ⋅  ⋅  1
 1  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  1  1  ⋅
 ⋅  1  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  1  1
 ⋅  ⋅  1  ⋅  ⋅  1  ⋅  ⋅  ⋅  1
 ⋅  ⋅  ⋅  1  ⋅  1  1  ⋅  ⋅  ⋅
 ⋅  ⋅  ⋅  ⋅  1  ⋅  1  1  ⋅  ⋅

julia> res = check_spectrum_integrality(A);

julia> res.matrix
10×10 Matrix{Int64}:
 0  1  0  0  1  1  0  0  0  0
 1  0  1  0  0  0  1  0  0  0
 0  1  0  1  0  0  0  1  0  0
 0  0  1  0  1  0  0  0  1  0
 1  0  0  1  0  0  0  0  0  1
 1  0  0  0  0  0  0  1  1  0
 0  1  0  0  0  0  0  0  1  1
 0  0  1  0  0  1  0  0  0  1
 0  0  0  1  0  1  1  0  0  0
 0  0  0  0  1  0  1  1  0  0

julia> res.spectrum_integral
true

julia> res.multiplicities
OrderedCollections.OrderedDict{Int64, Int64} with 3 entries:
  3  => 1
  -2 => 4
  1  => 5

Notes

If an undirected graph with integer edge weights is {-1,0,1}-diagonalizable (or, more restrictively, {-1,1}-diagonalizable), then its Laplacian matrix has integer eigenvalues [JP25, p. 312]. Hence, validating Laplacian integrality serves as a useful screening step in this package's principal S-bandwidth minimization algorithm.

It is, perhaps, an odd choice to sort the eigenvalue/multiplicity pairs in this file rather than in s_bandwidth—after all, in the context of the overarching S-bandwidth algorithm, this ordering is only ever used to determine which eigenspaces are searched for S-bases first. However, the inclusion of check_spectrum_integrality in the public API motivates a consistent, natural ordering of the multiplicites map in each (potentially user-facing) SpectrumIntegralResult instance.

Of course, we make it a point to still sort multiplicities as desired (first by ascending multiplicity then by ascending eigenvalue) in s_bandwidth itself; this seems more robust than relying on check_spectrum_integrality to do so or even folding the logic into the SpectrumIntegralResult inner constructor.

SDiagonalizability.laplacian_spectra_01neg — Method

laplacian_spectra_01neg(L::AbstractMatrix{<:Integer})

Compute data on the {-1,0,1}-spectrum of some Laplacian matrix L.

TODO: Write here

Arguments

L::AbstractMatrix{<:Integer}: the Laplacian matrix on whose {-1,0,1}-spectrum we are to compute data.

Returns

::LaplacianSpectrum01Neg: a struct containing the following fields:

laplacian_matrix::Matrix{Int}: a (casted) copy of L, avoiding shared mutability.

diagonalizable_01neg::Bool: whether the Laplacian is {-1,0,1}-diagonalizable.

diagonalizable_1neg::Bool: whether the Laplacian is {-1,1}-diagonalizable.

eigspaces_01neg::Union{Nothing,OrderedDict{Int,_Eigenspace01Neg}}: TODO: Write here

Examples

TODO: Write here

Notes

TODO: Include notes on the relevant properties of each type of Laplacian matrix

SDiagonalizability.s_bandwidth — Method

s_bandwidth()

TODO: Write here. Just a placeholder for Documenter.jl as of right now.

References

[Fox09]

J. Fox. Lecture 19: The Petersen graph and Moore graphs. Lecture notes, MAT 307: Combinatorics, Massachusetts Institute of Technology (2009).

[JP25]

N. Johnston and S. Plosker. Laplacian {−1,0,1}- and {−1,1}-diagonalizable graphs. Linear Algebra and its Applications 704, 309–39 (2025).

[Joy15]

D. Joyce. Rotations and complex eigenvalues. Lecture notes, Math 130: Linear Algebra, Clark University (2015).